% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{add_active_count_tag}
\alias{add_active_count_tag}
\title{add active count tag}
\usage{
add_active_count_tag(vesalius_assay, norm)
}
\arguments{
\item{vesalius_assay}{a vesalius assay object}

\item{norm}{requested count matrix}
}
\value{
commented list with active embedding tag
}
\description{
add active count tag
}
