% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{add_active_tag}
\alias{add_active_tag}
\title{add active tag}
\usage{
add_active_tag(vesalius_assay, embed_mat, embed)
}
\arguments{
\item{vesalius_assay}{a vesalius assay object}

\item{embed_mat}{matrix containing embeddings}

\item{embed}{embedding used}
}
\value{
a embedding matrix with added tag that describes
which embedding was used to produce that embedding
}
\description{
add active tag
}
