% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_modification.R
\name{add_cells}
\alias{add_cells}
\title{add cells to vesalius assay}
\usage{
add_cells(vesalius_assay, cells, add_name = NULL, verbose = TRUE)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}

\item{cells}{data.frame containing cell coordinates and metadata}

\item{add_name}{character string defining trial name for added cells}

\item{verbose}{logical - progress message output}
}
\value{
vesalius_assay object with added cells
}
\description{
add cells to vesalius assay
}
