% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_modification.R
\name{add_counts}
\alias{add_counts}
\title{Add counts to vesalius assay
Adding custom count matrix to a vesalius assay object.}
\usage{
add_counts(
  vesalius_assay,
  counts,
  raw_counts = NULL,
  count_type = "custom_counts",
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{vesalius_assay}{a vesalius assay object}

\item{counts}{matrix or sparse matrix containing normalised counts}

\item{raw_counts}{matrix or sparse matrix containing raw counts}

\item{count_type}{character string defining the name of the count matrix
being added.}

\item{force}{logical indicating if count matrix provided should also be used
as raw count matrix.}

\item{verbose}{logical indicating if progress messages should be outputed.}
}
\value{
a vesalius_assay object
}
\description{
Add counts to vesalius assay
Adding custom count matrix to a vesalius assay object.
}
\details{
In some case, you might wish to use your own normalisation method.
In this case, you can add your own matrix and specify the name you want to 
give to that count matrix. 

Differential gene expression tools
such as DESeq2 and edgeR require raw counts. As such, we recommend providing
raw counts as well. If you do not have raw counts, or do not wish to provide
them, you can set the force argument to TRUE. This will force vesalius to
generate a copy of your count matrix and use this as "raw" count matrix.

Since coordinates need to be parsed to the vesalius_assay contructor,
this function will also compared the your count matrix to the 
coordinate data. It will trim the count matrix based on barcodes shared
between both.
}
