% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_modification.R
\name{add_embeddings}
\alias{add_embeddings}
\title{Add embeddings
Add custom embeddings to vesalius objects}
\usage{
add_embeddings(
  vesalius_assay,
  embeddings,
  embedding_type = "custom_embedding",
  verbose = TRUE
)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}

\item{embeddings}{a matrix containing embedding values (see details)}

\item{embedding_type}{character string to be used as embedding name.}

\item{verbose}{logical indicating if progress message should be outputed.}
}
\value{
a vesalius_assay object
}
\description{
Add embeddings
Add custom embeddings to vesalius objects
}
\details{
Vesalius objects accepts custom embedding values that will be
used to generate images.

The embedding matrix should be in the form of a matrix with columns being
the latent space dimensiions and rows representing the barcodes present
in the data set.

Rownames should also be present and should represent the barcode name. 
These rownames are use to match the latent space embedding to its tile
in the image.
}
