% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{adjust_counts}
\alias{adjust_counts}
\title{adjust count}
\usage{
adjust_counts(coordinates, counts, cores = 1)
}
\arguments{
\item{coordinates}{data frame containing coordinates after reducing 
resolution and compressing cooridnates}

\item{counts}{count matrix}

\item{cores}{numeric describing number of cores to be used}
}
\value{
a count matrix with adjusted count values
}
\description{
adjust counts after reducing the resolution of the image tensor.
}
\details{
This function will check the coordinate file to 
see if any barcodes have been merged together. If so,
the counts will be adjusted by taking the average count value accross 
all barcodes that have been merged together.
}
