% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{align_index}
\alias{align_index}
\title{assign coordinates to matched indices}
\usage{
align_index(matched_index, coord)
}
\arguments{
\item{matched_index}{data.frame containing matching pairs of 
coordinates}

\item{seed}{data.frame containing seed coordinates}

\item{query}{data.frame containing quert cooridates}

\item{verbose}{logical - should progress message be outputed to the 
console}
}
\value{
adjusted query coordinate data.frame where each point
receives the coordinates of its best matche in the seed.
}
\description{
assign coordinates to matched indices
}
