% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{assign_centers}
\alias{assign_centers}
\title{assign centroid values to active embedding}
\usage{
assign_centers(vesalius_assay, clusters, kcenters, dimensions)
}
\arguments{
\item{vesalius_assay}{a vesalius assy object}

\item{clusters}{data.frame containing cluster values}

\item{kcenters}{matrix containing centroid values for each dimension}

\item{dimensions}{vector (nummeric / int) describin which latent space
dimensiuons shouls be used.}
}
\value{
matrix for the active embedding usiong color segementation
}
\description{
assign centroid values to active embedding
}
