% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_to_RGB.R
\name{assignRGBtoPixel}
\alias{assignRGBtoPixel}
\title{Assigning the RGB colour code to its specific barcode location}
\usage{
assignRGBtoPixel(rgb, coordinates, na.rm = TRUE)
}
\arguments{
\item{rgb}{list of colour codes in the format describe in the Detail section.  Output of rgbPCA.}

\item{coordinates}{data frame containing x and y coordinates of barcodes}

\item{na.rm}{logical describing if NA should be removed}
}
\value{
a data frame with five columns: barcodes (barcode names), xcoord (x coordinates),
ycoord (y coordinates), R (Red colour channel),G (Green colour channel) ,and B (Blue Colour channel).
}
\description{
Assigning the RGB colour code to its specific barcode location
}
\details{
assignRGBtoPixel remaps the colour code to x/y coordinates. \code{rgb} is the output of \code{rgbPCA}.
As such, it should be a list with three different levels.
Level 1 : each list element is a different slice
Level 2 : each list element is R G B channel for that slice
Level 3 : numeric vectors containing colour code for each barcode
}
