% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_to_RGB.R
\name{assingRGBtoPixelQuickBlock}
\alias{assingRGBtoPixelQuickBlock}
\title{Assigning the RGB colour code to coordinates in lower resolution.}
\usage{
assingRGBtoPixelQuickBlock(
  rgb,
  coordinates,
  resolution = 200,
  drop = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{rgb}{list of colour codes in the format describe in the Detail section.  Output of rgbPCA.}

\item{coordinates}{data frame containing x and y coordinates of barcodes}

\item{resolution}{size of "image matrix" to be used. n by n matrix where n = resolution}

\item{drop}{logical describing if empty location should be removed.}

\item{na.rm}{logical describing if NA should be removed}
}
\value{
a data frame with five columns: barcodes (barcode names), xcoord (column index),
ycoord (row index), R (Red colour channel),G (Green colour channel) ,and B (Blue Colour channel).
}
\description{
Assigning the RGB colour code to coordinates in lower resolution.
}
\details{
assignRGBtoPixel remaps the colour code to x/y coordinates at lower resolution.
This means that beads/spots will be assigned to a "pixel" within a matrix then exported as a data frame.
The current version does not return assigned colours as array images.
\code{rgb} is the output of \code{rgbPCA}.
As such, it should be a list with three different levels.
Level 1 : each list element is a different slice
Level 2 : each list element is R G B channel for that slice
Level 3 : numeric vectors containing colour code for each barcode
}
