% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{build_vesalius_assay}
\alias{build_vesalius_assay}
\title{build vesalius assay object}
\usage{
build_vesalius_assay(
  coordinates,
  counts,
  assay = "spatial_omics",
  adjust_coordinates = "origin",
  verbose = TRUE
)
}
\arguments{
\item{coordinates}{data.frame in the format 
barcodes, x, y.}

\item{counts}{matrix, sparse matrix containing counts .}

\item{assay}{character vector containing names of the assays 
(see details).}

\item{adjust_coordinates}{character of one of the following
"origin" or "norm" (see details).}

\item{verbose}{logical indicating if progress message should be
outputed or not.}
}
\value{
A vesalius_assay objecy
}
\description{
build a simple vesalius assay object from single count matrix and spatial
coordinate pair.
}
\details{
Along side this input data, you can provide a name 
to your assay. If none are provided, 
Vesalius will generate a set of names based on 
the default assay name "spatial_omics".

You can decide if you want to adjust the coordinates to the 
origin i.e remove unnecessary space or normalise the coordinates.
Norm is not recommened at the moment. 

We highly recommend using \code{\link{vesaliusObjects}} instead!
}
\examples{
data(vesalius)
# Single assay object
ves <- build_vesalius_assay(coordinates, counts)
}
