% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{build_vesalius_object}
\alias{build_vesalius_object}
\title{build a vesalius object}
\usage{
build_vesalius_object(
  coordinates,
  counts,
  assays = "spatial_omics",
  adjust_coordinates = "origin",
  verbose = TRUE
)
}
\arguments{
\item{coordinates}{data.frame or list of data.frame in the format 
barcodes, x, y (see details).}

\item{counts}{matrix, sparse matrix or list of matrices or sparse 
matrices containing counts (see details).}

\item{assays}{character vector containing names of the assays 
(see details).}

\item{adjust_coordinates}{character of one of the following
"origin" or "norm" (see details).}

\item{verbose}{logical indicating if progress message should be
outputed or not.}
}
\value{
A vesaliusObject
}
\description{
build a vesalius object from coordinate files and counts matrices
}
\details{
You can build a vesaliusObject by either supplying 
a single matrix/ coordinate combination or by providing a list
containing multiple matrices and coordinates. Please note that 
both list should be the same length. 

Along side this input data, you can provide a list of names 
to your assays. If none are provided or if not enough names
are provided, Vesalius will generate a set of names based on 
the default assay name "spatial_omics". Please ensure that 
the name you provide are unique to each assay.

You can decide if you want to adjust the coordinates to the 
origin i.e remove unnecessary space or normalise the coordinates.
Norm is not recommened at the moment.
}
\examples{
data(vesalius)
# Single assay object
ves <- build_vesalius_object(coordinates, counts)
# Building multi assay object
coord <- list(coordinates, coordinates)
co <- list(counts, counts)
assay_names <- c("my_assay", "my_second_assay")
ves <- build_vesalius_object(coord, co, assay_names)
}
