% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calculate_scale}
\alias{calculate_scale}
\title{calculate scale of assay}
\usage{
calculate_scale(coordinates, q = 0.999)
}
\arguments{
\item{coordinates}{Spatial coordinates as data frame}
}
\value{
Single float
}
\description{
calculate scale of assay
}
\details{
Calculate the average distance between spots/beads/indeces
}
