% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_barcodes}
\alias{check_barcodes}
\title{checking overlap between barcodes in counts and coordinates}
\usage{
check_barcodes(mat_barcodes, coordinates, throw = TRUE)
}
\arguments{
\item{mat_barcodes}{character vector containing barcode names in matrix
(count matrix or embedding matrix)}

\item{coordinates}{character vector containing barcode names in tile
data frame}

\item{throw}{logical if warning should be thrown or not}
}
\value{
overlapping location between counts and coordinates
}
\description{
checking overlap between barcodes in counts and coordinates
}
\details{
Will throw in a warning if the overlap is not perfect.
}
