% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_cell_labels}
\alias{check_cell_labels}
\title{check cell labels}
\usage{
check_cell_labels(vesalius_assay, trial = NULL)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}

\item{trial}{trial name}
}
\value{
cell labels data frame
}
\description{
check cell labels
}
