% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_cells}
\alias{check_cells}
\title{check if provided cells are contained withing provided territory}
\usage{
check_cells(territory_barcodes, ter, cell_barcodes, verbose)
}
\arguments{
\item{ter}{string selected territories in the form of chacater string}

\item{cell_barcodes}{character vector containing barcodes of cells
of interest}

\item{verbose}{logical if progress message should be outputed}

\item{territoty_barcodes}{character vector containing spatial barcodes
of territories}
}
\value{
charcater vector of common barcodes between territory 
barcodes and cell barcodes.
}
\description{
check if provided cells are contained withing provided territory
}
