% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_coordinates}
\alias{check_coordinates}
\title{check if coordinates are of the correct type and format and
adjust coordinate value to remove white edge space}
\usage{
check_coordinates(coordinates, assay, layer, verbose)
}
\arguments{
\item{coordinates}{coordinate data}

\item{assay}{string - assay name}

\item{layer}{int [1, Inf[ - depth layer}

\item{verbose}{logical if progress message should be printed}
}
\value{
coordinates data.frame or error
}
\description{
check if coordinates are of the correct type and format and
adjust coordinate value to remove white edge space
}
\details{
adjusts coordinates by either snapping coordinates to origin 
or min max normalisation of coordinates. Might add polar for future tests.
}
