% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_cost_contribution}
\alias{check_cost_contribution}
\title{check cost contribution}
\usage{
check_cost_contribution(vesalius_assay)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}
}
\value{
cost contribution data
}
\description{
check cost contribution
}
