% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_cost_matrix_validity}
\alias{check_cost_matrix_validity}
\title{check cost matrix validity}
\usage{
check_cost_matrix_validity(custom_cost)
}
\arguments{
\item{custom_cost}{custom cost matrix}
}
\value{
validated cost matrix
}
\description{
check cost matrix validity
}
