% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_cost_validity}
\alias{check_cost_validity}
\title{check cost validity}
\usage{
check_cost_validity(
  cost,
  seed_assay,
  seed_signal,
  query_assay,
  query_signal,
  use_cost,
  verbose = TRUE
)
}
\arguments{
\item{cost}{cost matrix}

\item{seed_assay}{seed assay}

\item{seed_signal}{seed signal}

\item{query_assay}{query assay}

\item{query_signal}{query signal}

\item{use_cost}{cost types to use}

\item{verbose}{logical - progress messages}
}
\value{
validated cost information
}
\description{
check cost validity
}
