% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_deg_method}
\alias{check_deg_method}
\title{check if DEG method is one of the available options}
\usage{
check_deg_method(method)
}
\arguments{
\item{method}{string - DEG method}
}
\value{
DEG method or error
}
\description{
check if DEG method is one of the available options
}
