% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_embed_methods}
\alias{check_embed_methods}
\title{check if embedding method is present in availble options}
\usage{
check_embed_methods(embed, n_counts)
}
\arguments{
\item{norm}{string - embedding method parsed to function}
}
\value{
embedding method string or error
}
\description{
check if embedding method is present in availble options
}
