% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_embedding_selection}
\alias{check_embedding_selection}
\title{check embedding selection}
\usage{
check_embedding_selection(vesalius_assay, embed, dims)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay}

\item{embed}{string embedding selection choice}

\item{dims}{integer vector containing embedding dimension to extract}
}
\value{
embedding data frame
}
\description{
check embedding selection
}
\details{
we want to check if the embedding that the user requests 
is present in the assay. If not return error. If more than one 
with that name return last entry of that name and warning. Default is last 
that will just take the last embedding created which should be stored in
the active slot in the vesalius_Assay object.
We also want to be able to select which dimensions we want to return.
We make sure that those dimensions can be extracted from the embedding 
data frame.
}
