% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_feature_integration}
\alias{check_feature_integration}
\title{check feature integration validity}
\usage{
check_feature_integration(signal, intergrated)
}
\arguments{
\item{signal}{signal type}

\item{intergrated}{integrated object}
}
\value{
validated features
}
\description{
check feature integration validity
}
