% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_features}
\alias{check_features}
\title{check features}
\usage{
check_features(counts)

check_features(counts)
}
\arguments{
\item{counts}{a seurat object}
}
\value{
a list of features

a list of features
}
\description{
check features

check features
}
\details{
If the user use custom counts, we want to check
if variable features have been computed or not. Dim reduction
requires a list if features toi be provided or variable features
to be computed.

If the user use custom counts, we want to check
if variable features have been computed or not. Dim reduction
requires a list if features toi be provided or variable features
to be computed.
}
