% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_for_unmatched}
\alias{check_for_unmatched}
\title{check for unmatched entries}
\usage{
check_for_unmatched(matched)
}
\arguments{
\item{matched}{matched data frame}
}
\value{
filtered matched data frame
}
\description{
check for unmatched entries
}
