% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_for_zero_counts}
\alias{check_for_zero_counts}
\title{check_for_zero_counts 
check which barcodes contain all 0 counts and filter}
\usage{
check_for_zero_counts(counts)
}
\arguments{
\item{counts}{count matrix}
}
\value{
trimmed count matrix
}
\description{
check_for_zero_counts 
check which barcodes contain all 0 counts and filter
}
