% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_gene_overlap}
\alias{check_gene_overlap}
\title{check gene overlap}
\usage{
check_gene_overlap(signal)
}
\arguments{
\item{signal}{signal matrix}
}
\value{
genes with non-zero expression
}
\description{
check gene overlap
}
