% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_genes}
\alias{check_genes}
\title{check genes validity}
\usage{
check_genes(genes, counts)
}
\arguments{
\item{genes}{gene names}

\item{counts}{count matrix}
}
\value{
validated genes
}
\description{
check genes validity
}
