% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_group_value}
\alias{check_group_value}
\title{check if requested territory is in territory list}
\usage{
check_group_value(territories, group)
}
\arguments{
\item{territories}{data frame containing territories}

\item{group}{vector indicating which territories should be selected
from territory data frame.}
}
\value{
a vector of territories that are present in territory data frame
}
\description{
check if requested territory is in territory list
}
