% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_image_input}
\alias{check_image_input}
\title{check image input 
check if input image is in correct format}
\usage{
check_image_input(image)
}
\arguments{
\item{image}{input}
}
\value{
list of images
}
\description{
check image input 
check if input image is in correct format
}
