% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_inputs}
\alias{check_inputs}
\title{check input checks the validity of input data to build_vesalius_assay}
\usage{
check_inputs(counts, coordinates, assay, adjust_coordinates, verbose)
}
\arguments{
\item{counts}{count matrix}

\item{coordinates}{coordinate file}

\item{assay}{string with assay name}

\item{adjust_coordinates}{string describing how coordinates should be
adjusted (origin or norm)}

\item{verbose}{logical if progress message should be outputed.}
}
\value{
list containing checked counts, coordinates and assay
}
\description{
check input checks the validity of input data to build_vesalius_assay
}
