% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_isolation_method}
\alias{check_isolation_method}
\title{check territory isolation method is available}
\usage{
check_isolation_method(method)
}
\arguments{
\item{method}{string - territory isolation method}
}
\value{
territory isolation method
}
\description{
check territory isolation method is available
}
