% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_map_input}
\alias{check_map_input}
\title{check map input validity}
\usage{
check_map_input(map)
}
\arguments{
\item{map}{map data structure}
}
\value{
validated map
}
\description{
check map input validity
}
