% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_map_selection}
\alias{check_map_selection}
\title{check map selection validity}
\usage{
check_map_selection(vesalius_assay, by)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}

\item{by}{selection criteria}
}
\value{
validated map columns
}
\description{
check map selection validity
}
