% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_maps}
\alias{check_maps}
\title{check maps availability}
\usage{
check_maps(vesalius_assay)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}
}
\value{
map data
}
\description{
check maps availability
}
