% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_metric_trial}
\alias{check_metric_trial}
\title{check metric trial validity}
\usage{
check_metric_trial(vesalius_assay, trial)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}

\item{trial}{trial name}
}
\value{
validated trial data
}
\description{
check metric trial validity
}
