% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_min_spatial_index}
\alias{check_min_spatial_index}
\title{check number of spatial indices present}
\usage{
check_min_spatial_index(group, min_spatial_index, id)
}
\arguments{
\item{group}{count matrix}

\item{min_spatial_index}{numeric - min number of spatial indices 
that need to be present in group}

\item{id}{string - id of group}
}
\description{
check number of spatial indices present
}
