% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_norm}
\alias{check_norm}
\title{check if select norm method is an available option}
\usage{
check_norm(vesalius_asssay, norm_method, method = NULL, verbose = TRUE)
}
\arguments{
\item{norm_method}{string - selected normalisation parsed by user}

\item{method}{DEG method}

\item{verbose}{logical if progress message should be ouputed}

\item{vesalius_assay}{a vesalius_assay}
}
\value{
count matrix
}
\description{
check if select norm method is an available option
}
\details{
Here we check if the normalisation method has been computed in
the count slot. We also check if this is used in the context of
differential gene expression analysis. DESeq and edgeR require raw counts
so if the user parses a valid norm method but is running DEG with one
those methods, we ignore request and parse raw count instead.
}
