% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_norm_methods}
\alias{check_norm_methods}
\title{check if norm method is present in availble options}
\usage{
check_norm_methods(norm)
}
\arguments{
\item{norm}{string - norm method parsed to function}
}
\value{
norm method string or error
}
\description{
check if norm method is present in availble options
}
