% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_sample}
\alias{check_sample}
\title{check sample information in vesalius assay}
\usage{
check_sample(vesalius_assay, sample = FALSE)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}

\item{sample}{logical - whether to return sample information}
}
\value{
sample information or NULL
}
\description{
check sample information in vesalius assay
}
