% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_segments}
\alias{check_segments}
\title{check if segment selection is a valid option}
\usage{
check_segments(vesalius_asssay, trial = "last")
}
\arguments{
\item{trial}{string - trial selection parse by user}

\item{vesalius_assay}{a vesalius_assay object}
}
\value{
data frame contain selected trial
}
\description{
check if segment selection is a valid option
}
\details{
check if the trial selection exists in territory slot
Default is last that will take the last entry. This function
will also reformat to only include the necessay information.
}
