% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_signal}
\alias{check_signal}
\title{check signal validity}
\usage{
check_signal(assay, signal, type)
}
\arguments{
\item{assay}{assay object}

\item{signal}{signal type}

\item{type}{data type}
}
\value{
validated signal
}
\description{
check signal validity
}
