% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_smoothing_kernel}
\alias{check_smoothing_kernel}
\title{check if smoothing method is present in availble options}
\usage{
check_smoothing_kernel(method)
}
\arguments{
\item{method}{smoothing method selected by user}
}
\value{
smoothing method string or error
}
\description{
check if smoothing method is present in availble options
}
