% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_template_source}
\alias{check_template_source}
\title{check template source validity}
\usage{
check_template_source(vesalius_assay, from, dimensions)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}

\item{from}{source type}

\item{dimensions}{dimensions to check}
}
\value{
validated template information
}
\description{
check template source validity
}
