% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_tensor_compression}
\alias{check_tensor_compression}
\title{check tensor compression}
\usage{
check_tensor_compression(locs)
}
\arguments{
\item{locs}{rle output of coordinate compression}
}
\value{
rle out of coordinate compression
}
\description{
check tensor compression
}
\details{
Here we just want to check if the output is reasonable or not
Essentially if you end up with very little barcodes, this could mean
that the user compressed to much.
Set warning if compression is set to less 10% of total points
}
