% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_territory_trial}
\alias{check_territory_trial}
\title{check if territory selection is a valid option}
\usage{
check_territory_trial(vesalius_assay, trial)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}

\item{trial}{string - trial selection parse by user}
}
\value{
data frame contain selected trial
}
\description{
check if territory selection is a valid option
}
\details{
check if the trial selection exists in territory slot
Default is last that will take the last entry. This function
will also reformat to only include the necessay information.
}
