% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_tiles}
\alias{check_tiles}
\title{check tile integretiy}
\usage{
check_tiles(vesalius_asssay)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay}
}
\value{
tile data frame
}
\description{
check tile integretiy
}
\details{
place holder for in depth checks at the moment only return tiles
}
