% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{chunk}
\alias{chunk}
\title{split vector into chunks}
\usage{
chunk(x, n, l = NULL)
}
\arguments{
\item{x}{vector to split}

\item{n}{chunk size}

\item{l}{maximum number of chunks}
}
\value{
list of chunks
}
\description{
split vector into chunks
}
