% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{chunker}
\alias{chunker}
\title{chunker}
\usage{
chunker(df, cores)
}
\arguments{
\item{df}{data.frame to chunk list elements}

\item{cores}{numeric number of chunks}
}
\value{
list of data frame chunks
}
\description{
creates data frame chunks to parse to multiple cores
}
