% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{circular_xcorr}
\alias{circular_xcorr}
\title{circular cross correlation}
\usage{
circular_xcorr(seed, query)
}
\arguments{
\item{seed}{seed signal as vector}

\item{query}{query signal as vector}
}
\value{
circularised cross correlation
}
\description{
circular cross correlation
}
