% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{clean_trial}
\alias{clean_trial}
\title{removing the "make.unique" labels when original labels are needed}
\usage{
clean_trial(trial)
}
\arguments{
\item{trial}{character vector containing barcode names}
}
\value{
cleaned trial vector
}
\description{
removing the "make.unique" labels when original labels are needed
}
