% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_pixel.R
\name{closet_grid_point}
\alias{closet_grid_point}
\title{closest grid point}
\usage{
closet_grid_point(coordinates, grid)
}
\arguments{
\item{coordinates}{data frame with spatial indices}

\item{grid}{grid matrix containing coordinate paiur of grid point centers}
}
\value{
closest coordinates to grid mid point
}
\description{
closest grid point
}
