% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{compare_niches}
\alias{compare_niches}
\title{Wrapper for fast niche composition}
\usage{
compare_niches(seed, query)
}
\arguments{
\item{seed}{list - cell type list in seed}

\item{query}{list - cell type list in query}
}
\value{
matrix - jaccard index between nich composition in 
seed and query
}
\description{
Wrapper for fast niche composition
}
