% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Territory_identity.R
\name{compareCells}
\alias{compareCells}
\title{compareCells compute differential gene expression between selected
territories for single cell type.}
\usage{
compareCells(
  image,
  counts,
  cells,
  seed = NULL,
  query = NULL,
  method = "wilcox",
  logFC = 0.25,
  pval = 0.05,
  minPct = 0.05,
  minBar = 10,
  morphologyFactorSeed = 0,
  morphologyFactorQuery = 0,
  verbose = TRUE
)
}
\arguments{
\item{image}{a dataframe containing territory information}

\item{counts}{seurat object containing counts. Alternatively, matrix or
sparse matrix. Colnames as barcodes and rownames as genes.}

\item{cells}{character vector containing barcodes associated with cell type
of interest.}

\item{seed}{Integer or vector of integers describing territories to be
included in group 1 for differential gene expression analysis.}

\item{query}{Integer or vector of integers describing territories to be
included in group 2 for differential gene expression analysis. Default = NULL}

\item{method}{character describing the statistical test to use in order to
extract differantial gene expression (currently only wilcox and t.test)}

\item{logFC}{numeric describing minimum log fold change value for
differential gene expression. Default set at 0.25.}

\item{pval}{numeric for pval threshold. Default set at 0.05}

\item{minPct}{numeric defining the minimum percentage of cells that should
contain any given gene.}

\item{minBar}{integer defining minimum number of barcodes in a territory.}

\item{morphologyFactorSeed}{Integer or vector of integers describing growth
or shrink rate of seed territories}

\item{morphologyFactorQuery}{Integer or vector of integers describing growth
or shrink rate of query territories}

\item{verbose}{logical - progress message output}
}
\value{
A data.frame (tibble) containing differential gene expression as well
p.value,
logFC,
seedPct (percentage of cells containing gene in first group),
queryPct (percentage of cells containing gene in second group),
seedTerritory (territory used as group 1)
queryTerritory (territory used as group 2)
}
\description{
compareCells compute differential gene expression between selected
territories for single cell type.
}
\details{
compareCells provides a way to compare the expression of cells
between different territories. If you have a cell type annotations associated
to each of the barcodes present in your data, you can select your cell 
type of interest and parse the barcodes to the compareCell function. 

Vesalius will internally subset your data and only compare
barcodes present in your territories of interest for differential 
gene expressioon. 

The cells argument should be supplied as a vector of barcodes containing
all barcodes of the cell type of interest. Theoretically, you can supply
more than one cell type in this vector but Vesalius will treat them as single
cell type.
}
\examples{
\dontrun{
data(vesalius)
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
# One segmentation round
image <- iterativeSegmentation.array(image)
image <- isolateTerritories.array(image, minBar = 5)
# For the example we will take a random selection of barcodes
cells <- GetAssayData(vesalius, slot = "data")
cells <- sample(colnames(cells), 100, replace =F)
markers <- compareCells(image, vesalius,cells seed = 1, query = 2)
}
}
