% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Territory_identity.R
\name{compareLayers}
\alias{compareLayers}
\title{compareLayers computes differential gene expression between territory layers}
\usage{
compareLayers(
  layers,
  counts,
  l1 = NULL,
  l2 = NULL,
  method = "wilcox",
  logFC = 0.25,
  pval = 0.05,
  minPct = 0.05,
  minBar = 10,
  verbose = TRUE
)
}
\arguments{
\item{layers}{data.frame containing layered territory information
(See \code{layerTerritory.edge} and \code{layerTerritory.concave})}

\item{counts}{seurat object containing counts. Alternatively, matrix or
sparse matrix. Colnames as barcodes and rownames as genes.}

\item{l1}{integer or vector of integers indicating layers to be contained in
group 1. Default NULL - If NULL will compare all layers independently}

\item{l2}{integer or vector of integers indicating layers to be contained in
group 2. If NULL will compare all layers independently}

\item{method}{character describing the statistical test to use in order to
extract differential gene expression (currently only wilcox and t.test)}

\item{logFC}{numeric describing minimum log fold change value for
differential gene expression. Default set at 0.25.}

\item{pval}{numeric for pval threshold. Default set at 0.05}

\item{minPct}{numeric defining the minimum percentage of cells that should
contain any given gene.}

\item{minBar}{integer defining minimum number of barcodes in a territory.}

\item{verbose}{logical - progress message output.}
}
\value{
A data.frame (tibble) containing differential gene expression as well
p.value,
logFC,
seedPct (percentage of cells containing gene in first group),
queryPct (percentage of cells containing gene in second group),
seedTerritory (territory used as group 1)
queryTerritory (territory used as group 2)
}
\description{
compareLayers computes differential gene expression between territory layers
}
\details{
Territory isolation provides a convenient way to investigate
the finer detail of anatomy and cellular spatial distribution. Image
representation of territories provides a convenient way to manipulate
territories such as dividing a territory into layers. Vesalius provides two
method to layer a territory : \code{layerTerritory.edge} and
\code{layerTerritory.concave}. The output of both function returns a data
frame with layer information. The \code{compareLayers} compares a layer or
layer sets between each other. If both l1 and l2 layers are left to NULL,
each layer will be compared to each other layer on at a time.
To compare one layer to all other layers, specific layer ID's should be
specified.
}
\examples{
\dontrun{
data("vesalius")
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
# One segmentation round
image <- iterativeSegmentation.array(image)
image <- isolateTerritories.array(image, minBar = 5)
layer <- layerTerritory.edge(image, seedTerritory = 1)
layerComp <- compareLayers(layer, vesalius, l1 = 1, l2 = 2)
}
}
