% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{compute_cell_cost}
\alias{compute_cell_cost}
\title{compute cell matching cost wrapper}
\usage{
compute_cell_cost(seed_signal, query_signal, i, verbose)
}
\arguments{
\item{seed_signal}{count matrix taken from seed data set}

\item{query_signal}{count matrix taken from query data set}

\item{i}{numeric - iteration value in query list}

\item{verbose}{logical - should progress message be outputed}
}
\value{
cost matrix with nrow = # cells in query and 
ncol = # cells in seed
}
\description{
compute cell matching cost wrapper
}
