% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territory_identity_and_markers.R
\name{compute_effect_size}
\alias{compute_effect_size}
\title{compute_effect_size}
\usage{
compute_effect_size(pval, seed, query)
}
\arguments{
\item{pval}{pvalue for a given gene}

\item{seed}{number of cells in seed}

\item{query}{number of cells in query}
}
\value{
efect size estimate for an unbalenced 
power analysis.
importFrom pwr pwr.2p2n.test
}
\description{
compute_effect_size
}
\details{
If pval is 0 we convert that to a very small number
pwr does not take 0 as input values.
DESeq might reutn NA pvals - just skip the effect size and 
return NA
}
