% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{compute_nearest_neighbor_graph}
\alias{compute_nearest_neighbor_graph}
\title{compute and greate nearest neighbor graph}
\usage{
compute_nearest_neighbor_graph(embeddings, k = 20, cores = 1)
}
\arguments{
\item{embeddings}{embedding matrix}

\item{k}{numeric describing number of nearest neighbors}

\item{cores}{numeric for number of cores to be used}
}
\value{
igraph object
}
\description{
compute and greate nearest neighbor graph
}
