% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{compute_neighbor_cost}
\alias{compute_neighbor_cost}
\title{compute cell neighborhood matching cost}
\usage{
compute_neighbor_cost(
  cost_mat,
  seed,
  query,
  seed_signal,
  query_signal,
  k,
  i,
  verbose
)
}
\arguments{
\item{seed}{data.frame of coordinates for seed}

\item{query}{data.frame of coordinates for query}

\item{seed_signal}{matrix - count matrix for seed}

\item{query_signal}{matrix - count matrix for query}

\item{k}{integer - how many neighbors should be considered as 
part of the local neighborhood}

\item{i}{numeric - iteration value in query list}

\item{verbose}{logical - should progress message be outputed}

\item{cost_matrix}{matrix containing assignement cost 
from each cell in query to each cell in seed (rows x cols)}
}
\value{
cost matrix with nrow = # cells in query and 
ncol = # cells in seed
}
\description{
compute cell neighborhood matching cost
}
