% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{compute_territory_area}
\alias{compute_territory_area}
\title{Calculate the total area of a territory}
\usage{
compute_territory_area(
  vesalius_assay,
  territory = NULL,
  trial = "last",
  use_rescaled = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}

\item{territory}{numeric/character/vector describing wich terriory(ies)
that will be use to compute area.}

\item{trial}{which territory trial should be used}

\item{use_rescaled}{logical - which value should be use for}
}
\value{
territory area
}
\description{
Calculate the total area of a territory
}
\details{
Compute the total area of a territory
}
