% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{concat_cost}
\alias{concat_cost}
\title{concat cost - pairwise sum of score complement}
\usage{
concat_cost(cost, use_cost, recip = TRUE)
}
\arguments{
\item{cost}{list - named list contained score matrices}

\item{use_cost}{character - which cost matrices to use}

\item{recip}{logical - use reciprocal matrix}
}
\value{
list with cost matrices
}
\description{
concat cost - pairwise sum of score complement
}
