% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{concat_embed}
\alias{concat_embed}
\title{create new embedding from jointly measure spatial omics}
\usage{
concat_embed(seed, query, dimensions, norm_method, dim_reduction, signal)
}
\arguments{
\item{seed}{vesalius_assay object of the first modality}

\item{query}{vesalius_assay object of the second modality}

\item{dimensions}{int - number of gray scale images to create}

\item{norm_method}{string describing which normalisation 
method to use. One of the following "log_norm", "SCT", "TFIDF", "raw"}

\item{dim_reduction}{string describing which dimensionality
reduction method should be used. One of the following:
"PCA", "PCA_L", "UMAP", "LSI", "LSI_UMAP"}
}
\value{
embedding matrix used as grey scale image stack
}
\description{
create new embedding from jointly measure spatial omics
}
\details{
Create new latent space using feature from both modalities
Creates a new feature matrix than in nromalized and converted to latent
space image stack.
}
