% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{concat_matches}
\alias{concat_matches}
\title{merging batch matches together}
\usage{
concat_matches(matched_indices)
}
\arguments{
\item{matched_indices}{list containing matched points in each batch}
}
\value{
data.frame with matched points
}
\description{
merging batch matches together
}
