% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{connected_points}
\alias{connected_points}
\title{create path from neighboring points}
\usage{
connected_points(trial)
}
\arguments{
\item{trail}{data frame containign x and y path}
}
\value{
order x y coordinates
}
\description{
create path from neighboring points
}
