% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{convexify}
\alias{convexify}
\title{convexify}
\usage{
convexify(xside, yside, indx, indy)
}
\arguments{
\item{xside}{vector of x coordinates}

\item{yside}{vector of y coordinates}

\item{indx}{central point x coordinate}

\item{indy}{central point y coordinate}
}
\value{
a data frame of ordered x and y coordinates.
}
\description{
order coordinates based on their angle around a central point
}
\details{
For rasterisation, the shape of the polygon must be as
convex as possible. To ensure that all points are in some sense 
in a convex form we order them based on their polar coordinate 
angle.
}
