% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{coordinate_fft}
\alias{coordinate_fft}
\title{converting territory path to FFT}
\usage{
coordinate_fft(path, partials, ranges)
}
\arguments{
\item{path}{data frame containing ordered x and y coordiates of a territory}

\item{partials}{int - number of FFT partials to return}

\item{ranges}{min and max range across coordinates}
}
\description{
converting territory path to FFT
}
