% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coordinates counts}
\alias{coordinates counts}
\title{Spatial coordinates and count matrix for vesalius}
\format{
coordinate data frame and sparse count matrix
\describe{
 \item{coordinates}{spatial coodinates of Puck_200815 taken from slide-seV2}
 \item{counts}{gene counts of Puck_200815 taken from slide-seV2}
}
}
\source{
\url{https://singlecell.broadinstitute.org/single_cell/study/SCP948/robust-decomposition-of-cell-type-mixtures-in-spatial-transcriptomics#study-download}
}
\usage{
data(vesalius)
}
\description{
Spatial coordinates and count matrix for vesalius
}
\keyword{datasets}
