% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_alpha}
\alias{create_alpha}
\title{create alpha value if territories need to be highlighted}
\usage{
create_alpha(territories, highlight, alpha)
}
\arguments{
\item{territories}{vesalius territories taken from a vesalius_assay}

\item{highlight}{numeric vector describing which territories should 
be highlighted}

\item{alpha}{tranaparent factor}
}
\value{
vector of alpha values
}
\description{
create alpha value if territories need to be highlighted
}
\details{
If highlight is null, will return the same alpha values 
for all territories
}
