% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_palette}
\alias{create_palette}
\title{create color palette from predefine scheme}
\usage{
create_palette(territories, randomise)
}
\arguments{
\item{territories}{vesalius territories taken from a vesalius_assay}

\item{randomise}{logical describing if colour palette should be 
randomised.}
}
\value{
color vector
}
\description{
create color palette from predefine scheme
}
\details{
We use a predefined palette that use colour blind friendly
base colours. We generate a color palette based on the number of 
territories present. If required the colours will be randomly assinged
to each territory. Note that as the territory plot 
return a ggplot object, you can easily override the color scheme.
}
