% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{create_trial_tag}
\alias{create_trial_tag}
\title{create a trail tag name}
\usage{
create_trial_tag(trials, tag)
}
\arguments{
\item{trials}{character vector containing names of trials 
that have already been computed}

\item{tag}{character string describing which trial tag to add}
}
\description{
create a trail tag name
}
