% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{depth_neighborhood}
\alias{depth_neighborhood}
\title{Graph depth method based method to select niche}
\usage{
depth_neighborhood(coord, depth)
}
\arguments{
\item{coord}{data.frame - coordinates of spatial indices in assay}

\item{depth}{int - graph path depth to define neighborhood 
0 = self, 1 = direct neigbors, 2 = neighbors of neighbors, etc}
}
\value{
list containing barcodes of  neighbors for each 
spatial index.
}
\description{
Graph depth method based method to select niche
}
