% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{dispatch_cost_groups}
\alias{dispatch_cost_groups}
\title{dispatch barcodes to subset cost for match clustering}
\usage{
dispatch_cost_groups(
  vesalius_assay,
  cost,
  trial = NULL,
  group_identity = NULL,
  ref_cells = NULL,
  query_cells = NULL
)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object post cell mapping}

\item{group_identity}{character - name of column containing group
to be used for cluster (i.e. territories, segments, layers etc)}

\item{cell_label}{character - name of column containing cell names
if the clustering is to be done by cell types only}
}
\value{
vector of strings (barcodes)
}
\description{
dispatch barcodes to subset cost for match clustering
}
