% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_dispatch.R
\name{dispatch_deg_group}
\alias{dispatch_deg_group}
\title{dispatch barcodes to seed and query groups}
\usage{
dispatch_deg_group(ter, seed, query, cells, verbose)
}
\arguments{
\item{ter}{territories data frame from territoires slot in
vesalius_assay object}

\item{seed}{interger vector indicating which territories should be included
in seed group}

\item{query}{interger vector indicating which territories should be included
in query group}

\item{cells}{cell barcodes}

\item{verbose}{logical if progress messages should be outputed}
}
\value{
list with seed group and seed id as well as query group and query id
}
\description{
dispatch barcodes to seed and query groups
}
\details{
This function generates groups for DEG analysis. It creates 
different groups depending on what is being parse to both seed and query. 
This could probably be cleaned up and simplified. Always need to return 
a list though.
}
