% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_dispatch.R
\name{dispatch_sample}
\alias{dispatch_sample}
\title{dispatch barcodes associted with a specific territory in a specific samples}
\usage{
dispatch_sample(territory_barcodes, ter, sample)
}
\arguments{
\item{territory_barcodes}{character vector barcodes associated with territory}

\item{ter}{chracter territory id}

\item{sample}{character vector associeted with sample}
}
\value{
character vector of barcodes at the intersection of territory and sample
}
\description{
dispatch barcodes associted with a specific territory in a specific samples
}
