% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{distance_pooling}
\alias{distance_pooling}
\title{distance pooling beads of colour segment into seperate territories}
\usage{
distance_pooling(img, capture_radius, min_spatial_index)
}
\arguments{
\item{img}{data frame contain all barcodes of a single sgement}

\item{capture_radius}{numeric proportion of max distance between beads
to use as distance threshold between beads}

\item{min_spatial_index}{numeric minimum number of beads that should
be contained in a territory.}
}
\description{
distance pooling beads of colour segment into seperate territories
}
\details{
Beads that are too far away or bead cluster that are below
the minimum number of spatial indices will all be pooled under the
isolated label. Note that this label is used across color segments.
}
