% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{divide_and_conquer}
\alias{divide_and_conquer}
\title{splitting cost matrix into batches for batch processing}
\usage{
divide_and_conquer(cost_mat, batch_size)
}
\arguments{
\item{cost_mat}{matrix}

\item{batch_size}{integer - size of each batch}
}
\value{
list containing cost matrix batches
}
\description{
splitting cost matrix into batches for batch processing
}
