% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{.thresholdMlMask}
\alias{.thresholdMlMask}
\title{Segment territories by iterative threshold-based segmentation.}
\usage{
.thresholdMlMask(
  img,
  tiles,
  elementSize = 5,
  maskType = "double",
  distThresh = 0,
  sizeThresh = 0
)
}
\arguments{
\item{img}{a cimg object}

\item{tiles}{from vesalius object}

\item{elementSize}{an integer describing the size of the element for erosion and dilation.}

\item{maskType}{a string describing mask type to use. "pos" segments territories with pixel values higher than the median, "neg" lower than the median and "double" on both sides. It is recommended to use "double".}

\item{distThresh}{a float describing the minimum mean pixel value distance of a territory from the image median. Should only be used in combination with quantile normalized embeddings.}

\item{sizeThresh}{an integer describing the minimum size of a territory in pixels.}
}
\value{
Territories: as a list of a list of bar codes.
}
\description{
Segment territories by iterative threshold-based segmentation.
}
