% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{.thresholdMlSmooth}
\alias{.thresholdMlSmooth}
\title{Iterative threshold-based segmentation.}
\usage{
.thresholdMlSmooth(
  img,
  maskType = "double",
  elementSize = 5,
  distThresh = 0,
  sizeThresh = 0
)
}
\arguments{
\item{img}{a cimg object}

\item{maskType}{a string describing mask type to use. "pos" segments territories with pixel values higher than the median, "neg" lower than the median and "double" on both sides.}

\item{elementSize}{an integer describing the size of the element for erosion and dilation.}

\item{distThresh}{a float describing the minimum mean pixel value distance of a territory from the image median. Should only be used in combination with quantile normalized embeddings}

\item{sizeThresh}{an integer describing the minimum size of a territory in pixels.}

\item{dev}{logical: return segmented image before assigning labels to territories. Only for development purposes.}
}
\value{
a cimg object. A segmented image with territories labeled continuously with positive integers and the background labeled as 0.
}
\description{
Iterative threshold-based segmentation.
}
