% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{embed_latent_space}
\alias{embed_latent_space}
\title{embed latent space}
\usage{
embed_latent_space(
  counts,
  assay,
  dim_reduction,
  dimensions,
  remove_lsi_1,
  verbose
)
}
\arguments{
\item{counts}{Seurat object containing counts (generally normalised)}

\item{assay}{charcter string of the assay being used}

\item{dim_reduction}{dimensionality reduction method that will be used
Select from PCA, PCA_L, UMAP, LSI, LSI_UMAP}

\item{dimensions}{numeric for number of dimeniosn top retain after 
dimensionality reduction}

\item{remove_lsi_1}{logical if first dimension of LSI embedding should be 
removed (will soon be depreciated)}

\item{verbose}{logical if progress messages should be outputed or not}
}
\value{
data frame of normalised embedding values.
}
\description{
Embed latent space into grey color scale.
}
\details{
General method dispatch function for dim reduction methods
}
