% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{embed_nmf}
\alias{embed_nmf}
\title{embed nmf}
\usage{
embed_nmf(counts, dimensions, verbose = TRUE)
}
\arguments{
\item{counts}{Seurat object containing normalised counts}

\item{dimensions}{number dimension to retain from NMF}

\item{verbose}{logical if progress messages should be outputed}
}
\value{
normalised NMF embedding matrix
}
\description{
embed in grey scale using NMF embeddings
}
