% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{embed_pca}
\alias{embed_pca}
\title{embed PCA}
\usage{
embed_pca(counts, dimensions, features = NULL, verbose = TRUE)
}
\arguments{
\item{counts}{Seurat object containing normalised counts}

\item{dimensions}{number dimension to retain from PCA}

\item{features}{custom vector of features}

\item{verbose}{logical if progress messages should be outputed}
}
\value{
normalised PCA embedding matrix
}
\description{
embed in grey scale using PCA embeddings
}
