% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{embed_pcal}
\alias{embed_pcal}
\title{embed PCA loading values}
\usage{
embed_pcal(counts, dimensions, features = NULL, verbose = TRUE)
}
\arguments{
\item{counts}{Seurat object containing normalised counts}

\item{dimensions}{number dimension to retain from PCA}

\item{features}{custom vector of features}

\item{verbose}{logical if progress messages should be outputed}
}
\value{
normalised PCA loading matrix
}
\description{
embed in grey scale using PCA Loading value
}
\details{
This approach is a slightly different as it takes 
the loading value associted to each gene in a barcode and sums
the absolute value of each of those values. 
Once all genes in all barcodes have been summed,
we normalise the latent space and return the matrix.
}
