% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsolatingTerritories.R
\name{enhance}
\alias{enhance}
\title{Enhance colour contrast by equalizing the colour histogram}
\usage{
enhance(img)
}
\arguments{
\item{img}{data frame with barcodes, xcoord, ycoord, R, G, and B as coloumns}
}
\value{
data frame with barcodes, xcoord, ycoord, R, G, and B as coloumns with adjusted RGB values.
}
\description{
Enhance colour contrast by equalizing the colour histogram
}
\details{
Histogram equalisation ensure that colours are uniformly distributed between all possible colours
Conceptually, this is similar to flatening a curve.
}
