% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{equalize_histogram}
\alias{equalize_histogram}
\title{equalizeHistogram image enhancement via colour histogram equalization.}
\usage{
equalize_histogram(
  vesalius,
  dimensions = seq(1, 3),
  embedding = "last",
  type = "BalanceSimplest",
  N = 1,
  smax = 1,
  sleft = 1,
  sright = 1,
  lambda = 0.1,
  up = 100,
  down = 10,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{type}{character - histogram EQ type. Select from: BalanceSimplest,
EqualizePiecewise, SPE, EqualizeDP, EqualizeADP, ECDF (see details)}

\item{N}{numeric describing how each colour channel will be mapped back to
the image (Higher N = Higher greyscale contrast).
Used with EqualizePiecewise}

\item{smax}{numeric - upper limit if contrast stretching.
Used with EqualizePiecewise}

\item{sleft}{numeric - Range 0 - 100. Percentage of pixel to be saturated on
the left side of the histogram. Used with BalanceSimplest}

\item{sright}{numeric - Range 0 - 100. Percentage of pixel to be saturated on
the right side of the histogram. Used with BalanceSimplest}

\item{lambda}{numeric - strength of background correction.
Used with SPE (Screened Poisson Equation).}

\item{up}{numeric - color value threshold in the upper limit.
Used with EqualizeDP.}

\item{down}{numeric color value threshold in the lower limit.
Used with EqualizeDP.}

\item{verbose}{logical - progress message output.}

\item{image}{data.frame - Vesalius formatted data.frame (i.e. barcodes,
x,y,cc,value,...)}

\item{invert}{logical - If TRUE, colours will be inverted i.e. 1 - colorValue
(background set to 1 instead of 0).}
}
\value{
Returns a Vesalius data.frame with "barcodes","x","y","cc",
"value","tile",...
}
\description{
equalizeHistogram image enhancement via colour histogram equalization.
}
\details{
Histogram equalization ensures that image details are amplified.
In turn, territories may be extract with greater precision. We recommend
balancing the histogram prior to smoothing.

For further details on each method described here, please refer to
\href{imagerExtra Vignette}{https://cran.r-project.org/web/packages/imagerExtra/vignettes/gettingstarted.html}
}
\examples{
\dontrun{
data(vesalius)
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
image <- equalizeHistogram(image)
imagePlot(image)
}
}
