% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{equalize_image}
\alias{equalize_image}
\title{equalise image histogram}
\usage{
equalize_image(
  vesalius_assay,
  dimensions = seq(1, 3),
  embedding = "last",
  type = "BalanceSimplest",
  N = 1,
  smax = 1,
  sleft = 1,
  sright = 1,
  lambda = 0.1,
  up = 100,
  down = 10,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}

\item{dimensions}{numeric vector of latent space dimensions to use.}

\item{embedding}{character string describing which embedding should
be used.}

\item{type}{character - histogram EQ type. Select from: BalanceSimplest,
EqualizePiecewise, SPE, EqualizeDP, EqualizeADP, ECDF (see details)}

\item{N}{numeric describing how each colour channel will be mapped back to
the image (Higher N = Higher greyscale contrast).
Used with EqualizePiecewise}

\item{smax}{numeric - upper limit if contrast stretching.
Used with EqualizePiecewise}

\item{sleft}{numeric - Range 0 - 100. Percentage of pixel to be saturated on
the left side of the histogram. Used with BalanceSimplest}

\item{sright}{numeric - Range 0 - 100. Percentage of pixel to be saturated on
the right side of the histogram. Used with BalanceSimplest}

\item{lambda}{numeric - strength of background correction.
Used with SPE (Screened Poisson Equation).}

\item{up}{numeric - color value threshold in the upper limit.
Used with EqualizeDP.}

\item{down}{numeric color value threshold in the lower limit.
Used with EqualizeDP.}

\item{cores}{numeric number of cores to be used}

\item{verbose}{logical - progress message output.}
}
\value{
a vesalius_assay object
}
\description{
equalizeHistogram image enhancement via colour histogram equalization.
}
\details{
Histogram equalization ensures that image details are amplified.
In turn, territories may be extract with greater precision. We recommend
balancing the histogram prior to smoothing.

For further details on each method described here, please refer to
\href{imagerExtra Vignette}{https://cran.r-project.org/web/packages/imagerExtra/vignettes/gettingstarted.html}
}
\examples{
\dontrun{
data(vesalius)
# First we build a simple object
ves <- build_vesalius_object(coordinates, counts)
# We can do a simple run
ves <- build_vesalius_embeddings(ves)

# simple EQ
ves <- equalisz_image(ves, embedding = "PCA")
}
}
