% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Territory_identity.R
\name{extractAllMarkers}
\alias{extractAllMarkers}
\title{extractAllMarkers compute differential gene expression for each territory}
\usage{
extractAllMarkers(
  image,
  counts,
  method = "wilcox",
  logFC = 0.25,
  pval = 0.05,
  minPct = 0.05,
  minBar = 10,
  morphologyFactor = 0,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{image}{a dataframe containing territory information}

\item{counts}{seurat object containing counts. Alternatively, matrix or
sparse matrix. Colnames as barcodes and rownames as genes.}

\item{method}{character describing the statistical test to use in order to
extract differential gene expression (currently only wilcox and t.test)}

\item{logFC}{numeric describing minimum log fold change value for
differential gene expression. Default set at 0.25.}

\item{pval}{numeric for pval threshold. Default set at 0.05}

\item{minPct}{numeric defining the minimum percentage of cells that should
contain any given gene.}

\item{minBar}{integer defining minimum number of barcodes in a territory.}

\item{morphologyFactor}{Integer or vector of integers describing growth or
shrink rate of territories.}

\item{cores}{numeric describing the number of cores used for the analysis.}

\item{verbose}{logical - progress message output}
}
\value{
A data.frame (tibble) containing differential gene expression as well
p.value,
logFC,
seedPct (percentage of cells containing gene in first group),
queryPct (percentage of cells containing gene in second group),
seedTerritory (territory used as group 1)
queryTerritory (territory used as group 2)
}
\description{
extractAllMarkers compute differential gene expression for each territory
}
\details{
extractAllMarkers function compares a territory after any
morphological operator to all remaining barcodes.
If one of territory sets does not contain enough
barcodes, \code{extractMarkers} will return NULL. For example, if you dilate
a territory (let's say you want to grow each territory by 5 pixels -
morphologyFactor = 5),the seed territory will be dilated and compared to all
remaining barcodes that not part of this new dilated territory.

For further details on territory Morphology - refer to
\code{territoryMorphology}
}
\examples{
\dontrun{
data(vesalius)
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
# One segmentation round
image <- iterativeSegmentation.array(image)
image <- isolateTerritories.array(image, minBar = 5)
mark <- extractAllMarkers(image, vealius )

}
}
