% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsolatingTerritories.R
\name{extractTerritories}
\alias{extractTerritories}
\title{Extracting Territories from image data}
\usage{
extractTerritories(img, territories)
}
\arguments{
\item{img}{data frame with barcodes, xcoord, ycoord, R, G, B, cluster number, territory number within that cluster.}

\item{territories}{list of vectors (cluster and territory).}
}
\value{
data frame with barcodes, xcoord, ycoord, R, G, B, cluster number, territory number within that cluster where
combined territories will take the value of the first cluster/territory provided.
}
\description{
Extracting Territories from image data
}
\details{
In order to combine territories together, this function takes a list of territories that should be
combined into one. The list elements are vectors containing first the cluster number and then the territory
number within that cluster.
}
