% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{feature_dissim}
\alias{feature_dissim}
\title{compute the distance between seed and query signals}
\usage{
feature_dissim(seed, query)
}
\arguments{
\item{seed}{seed signal}

\item{query}{query signal}
}
\value{
matrix with query as rows and seed as colmuns
}
\description{
compute the distance between seed and query signals
}
