% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{feature_dist}
\alias{feature_dist}
\title{compute the distance between seed and query signals}
\usage{
feature_dist(seed, query, batch_size)
}
\arguments{
\item{seed}{seed signal}

\item{query}{query signal}

\item{batch_size}{int describing size of each batch parse to the loops}
}
\value{
matrix with query as rows and seed as colmuns
}
\description{
compute the distance between seed and query signals
}
\details{
Chunking cost and signal into smaller chunks to run the 
correlation score in paralell. There is room for improvement here.
First we could dispatch the longer list to future_lapply
but cannot know which one it is and we need to know that so we can 
subset the cost. 
Also the functions calls feature_cost which is a R wrapper for a 
c++ function (cost.cpp).
}
