% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{filter_assay}
\alias{filter_assay}
\title{filter vesalius assay by cells, territories, or genes}
\usage{
filter_assay(
  vesalius_assay,
  cells = NULL,
  territories = NULL,
  trial = "last",
  genes = NULL
)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}

\item{cells}{character vector of cell barcodes to keep}

\item{territories}{character vector of territory labels to keep}

\item{trial}{character string defining which territory trial to use}

\item{genes}{character vector of genes to keep}
}
\value{
filtered vesalius_assay object
}
\description{
filter vesalius assay by cells, territories, or genes
}
