% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{filter_cost}
\alias{filter_cost}
\title{remove scores that are below a certain threshold}
\usage{
filter_cost(cost, threshold = 0.3, use_cost)
}
\arguments{
\item{cost}{list - score list to be filtered}

\item{threshold}{numeric - score threshold [0,1]}

\item{use_cost}{- character - which score matrices to filter}
}
\value{
filtered cost list
}
\description{
remove scores that are below a certain threshold
}
