% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{filter_grid}
\alias{filter_grid}
\title{filter grid}
\usage{
filter_grid(coordinates, filter_grid)
}
\arguments{
\item{coordinates}{data frame containing barcodes, / y coordinates of 
each barcode.}

\item{filter_grid}{numeric describing size of the grid to use as proporiton 
of array size.}
}
\value{
a data.frame containing barcodes, x and y coordinates.
}
\description{
filtered stray barcodes/spatial indices.
}
\details{
we create a grid over the data an check which grid section contain
a number of barcodes lower than a quantile threshold. Those barcodes will 
be removed. As it stands, I am not satisfied with this function. I think 
it is too restrictive and will most likely not be applicable to highly 
standardised assay.
}
