% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metrics.R
\name{filter_maps}
\alias{filter_maps}
\title{remove scores that are below a certain threshold}
\usage{
filter_maps(mapped, threshold, allow_duplicates, verbose)
}
\arguments{
\item{threshold}{numeric - score threshold [0,1]}

\item{cost}{list - score list to be filtered}

\item{use_cost}{- character - which score matrices to filter}
}
\value{
filtered cost list
}
\description{
remove scores that are below a certain threshold
}
