% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{filter_tiles}
\alias{filter_tiles}
\title{filter tiles}
\usage{
filter_tiles(tesselation, coordinates, filter_threshold)
}
\arguments{
\item{tesselation}{data.frame output from the deldir function}

\item{coordinates}{data frame with original coordinates}

\item{filter_threshold}{numeric describing the quantile threshold value
to use for area filtering}
}
\value{
a list with 2 data frame. 1 with filtered tesselation results
2 filtered coordinate file.
}
\description{
filter tiles based on tile area and if they share an edge with the 
tesselation box
}
\details{
Here we want to filter based on the size of the tile
under the assumption that very large tiles are probably due to 
unpexted space. The issue is that if you don't apply this threshold,
subtle pixel patterns are lost wuhtin the sea of pixels. 
If the ueser really dont want to filter anything out then you don't
We will set a filter threshold pretty high though to make sure 
that it does get filter out as default.
}
