% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filterCountMatrix}
\alias{filterCountMatrix}
\title{filter count Matrix - removing low information beads}
\usage{
filterCountMatrix(slide, min.genes = 50, min.count = 100)
}
\arguments{
\item{slide}{count matrix with beads/barcodes as columns and genes as rows}

\item{min.genes}{minimum number of different genes a barcodes should have}

\item{min.count}{minimum number of counts per barcode}
}
\value{
A reduced version of the count matrix with barcodes as columns and
genes as rows
}
\description{
filter count Matrix - removing low information beads
}
