% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNN.R
\name{findNearestNeighbors}
\alias{findNearestNeighbors}
\title{Finding Nearest spatial neighbors}
\usage{
findNearestNeighbors(slide, k = 5, box = 0.05, sectors = 4, cores = 1)
}
\arguments{
\item{slide}{a slide containing the barcodes and associated x/y coordinates.}

\item{k}{the number of nearest neighbors to extract}

\item{box}{the size of the box used to find nearest neighbors.
Default is set at 0.05. This represents the proportion of total distance.}

\item{sectors}{the number of angle sectors used to find nearest neighbors.
Default is set at four.}

\item{cores}{number of cores used to find nearest neighbors}
}
\value{
A list of nearest neighbors. Each element in the list represents a single
    barcode used as center point. Each elements in the list is a data frame with neighbor barcodes
    x and y coordninates of said neighbors as well as distance and angle from center point.
}
\description{
Finding Nearest spatial neighbors
}
\details{
In order to ensure that the nearest neighbors are distributed in all
   directions, this function find all neighbors within a certain distance (box) and
   categorises all points based on their angle compared to the center point. These points are
   then sorted. The resulting points are the points that are both the closest
   but also found in all directions compared to a center point. If direction is not
   considered important, then set sectors to 1.
}
