% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsolatingTerritories.R
\name{findSubClusters}
\alias{findSubClusters}
\title{findSubCluster - Clustering of territories}
\usage{
findSubClusters(
  img,
  SO,
  by = c("cluster", "territory"),
  varF = 2000,
  npcs = 20,
  resolution = 0.5
)
}
\arguments{
\item{img}{data frame with barcodes, xcoord, ycoord, R, G, B, cluster number, territory number within that cluster.}

\item{SO}{Seurat Object with full spatial assay}

\item{by}{character describing if subclustering should be caried out on colour cluster or territories.}

\item{varF}{integer describing number of variable features}

\item{npcs}{integer describing the number of Principle components to compute}

\item{resolution}{numeric describing granularity of clustering}
}
\value{
A list of lists. The first level represents each cluster of territory. For each cluster or territory,
the list element contains a Seurat Object with sub cluster and markers associated to each subcluster.
}
\description{
findSubCluster - Clustering of territories
}
\details{
The following function is a wrapper function that performs the standard Seurat analysis.
}
