% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_dispatch.R
\name{format_c_to_ves}
\alias{format_c_to_ves}
\title{convert cimg list to vesalius object embedding}
\usage{
format_c_to_ves(cimg, vesalius_assay, dims, embed = "last", verbose = TRUE)
}
\arguments{
\item{cimg}{cimg image list}

\item{vesalius_assay}{a vesalius_assay object}

\item{dims}{integer vector of embedding that need to be updated}

\item{embed}{character string which embedding should be updated}

\item{verbose}{logical should progress message be outputed}
}
\value{
a vesalius_Assay obejct
}
\description{
convert cimg list to vesalius object embedding
}
