% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_dispatch.R
\name{format_call}
\alias{format_call}
\title{format function call to list for log update}
\usage{
format_call(call)
}
\arguments{
\item{call}{a function call argument list}
}
\description{
format function call to list for log update
}
