% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_dispatch.R
\name{format_counts_for_deseq2}
\alias{format_counts_for_deseq2}
\title{format counts for DESeq}
\usage{
format_counts_for_deseq2(seed, query)
}
\arguments{
\item{seed}{seed count matrix}

\item{query}{query count matrix}
}
\value{
DESeq2 object
}
\description{
format counts for DESeq
}
\details{
Always adding a pseudocount of 1
}
