% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{format_counts_for_edger}
\alias{format_counts_for_edger}
\title{format counts for edgeR}
\usage{
format_counts_for_edger(seed, query)
}
\arguments{
\item{seed}{seed count matrix}

\item{query}{query count matrix}
}
\value{
DGEList object from edgeR
}
\description{
format counts for edgeR
}
