% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_dispatch.R
\name{format_counts_for_logit}
\alias{format_counts_for_logit}
\title{format counts for logistic regression}
\usage{
format_counts_for_logit(seed, query)
}
\arguments{
\item{seed}{seed count matrix}

\item{query}{query count matrix}
}
\value{
list of merged counts and meta data formated for logit
}
\description{
format counts for logistic regression
}
