% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{format_ves_to_c}
\alias{format_ves_to_c}
\title{convert vesalius_assay to cimg images}
\usage{
format_ves_to_c(vesalius_assay, dims, embed = "last", verbose = TRUE)
}
\arguments{
\item{vesalius_assay}{a vesalius_Assay object}

\item{dims}{integer vector indicating the number of dimensions to select from
embeddings.}

\item{embed}{character indicating which embedding should be selected.
Default uses last embedding produced}

\item{verbose}{logical if progress message should be outputed}
}
\value{
list of cimg images
}
\description{
convert vesalius_assay to cimg images
}
