% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{generate_image_template}
\alias{generate_image_template}
\title{create a registration template image}
\usage{
generate_image_template(
  vesalius_assay,
  from = "Territory",
  dimensions = seq(1:3)
)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}

\item{from}{character string defining what should be used to create
the image template}

\item{dimensions}{numeric vector defining which dimensions should be used
if calling en embedding}
}
\value{
a cimg object
}
\description{
create a registration template image
}
\details{
creating a template gray image that we will use for image registration
The point here is to make sure that we have something can use to compared the
the image to a microscopy image
}
