% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{generate_slic_graph}
\alias{generate_slic_graph}
\title{importFrom kohonen som map scale somgrid}
\usage{
generate_slic_graph(spix, k = "auto")
}
\description{
importFrom kohonen som map scale somgrid
}
