% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spix.R
\name{generate_spix}
\alias{generate_spix}
\title{Generate super pixels from ST}
\usage{
generate_spix(
  vesalius_assay,
  dimensions = seq(1, 3),
  embedding = "last",
  method = "kmeans",
  col_resolution = 10,
  compactness = 1,
  scaling = 0.5,
  threshold = 0.9,
  index_selection = "bubble",
  max_iter = 10000,
  verbose = TRUE
)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}

\item{dimensions}{numeric vector of latent space dimensions to use.}

\item{embedding}{character string describing which embedding should
be used.}

\item{method}{character string for which method should be used for
segmentation. Select from "slic", 
"leiden_slic","louvain_slic"}

\item{col_resolution}{numeric colour resolution used for segmentation. 
(see details)}

\item{compactness}{numeric - factor defining super pixel compaction.}

\item{scaling}{numeric - scaling image ration during super pixel 
segmentation.}

\item{threshold}{numeric [0,1] - correlation threshold between
nearest neighbors when generating segments from super pixels.}

\item{index_selection}{character - method for selecting initial indices}

\item{max_iter}{int - max number of kmean iterations for slic segments}

\item{verbose}{logical - progress message output.}
}
\value{
a vesalius_assay object
}
\description{
Generate super pixels from ST
}
