% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{generate_territory_graph}
\alias{generate_territory_graph}
\title{territory graph}
\usage{
generate_territory_graph(territories, k)
}
\arguments{
\item{territories}{territory coordinates}

\item{k}{number of neigbors to create graph}
}
\value{
nearest neighbors matrix
}
\description{
territory graph
}
