% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{get_active_count_tag}
\alias{get_active_count_tag}
\title{get last count matrix used}
\usage{
get_active_count_tag(vesalius_assay)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}
}
\value{
character string with name of last embedding used
}
\description{
get last count matrix used
}
