% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{get_assay_names}
\alias{get_assay_names}
\title{get assay names from vesalius_assay}
\usage{
get_assay_names(vesalius_assay)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay}
}
\value{
assay names
}
\description{
get assay names from vesalius_assay
}
