% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{get_coordinates}
\alias{get_coordinates}
\title{get coordinates from vesalius_assay}
\usage{
get_coordinates(vesalius_assay, original = FALSE)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay}

\item{original}{logical - return original coordinates}
}
\value{
coordinate data frame
}
\description{
get coordinates from vesalius_assay
}
