% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{get_counts}
\alias{get_counts}
\title{get counts from vesalius_assay}
\usage{
get_counts(vesalius_assay, type = "raw")
}
\arguments{
\item{vesalius_assay}{a vesalius_assay}

\item{type}{character string for which count matrix to
retrieve.}
}
\value{
count matrix
}
\description{
get counts from vesalius_assay
}
