% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territory_identity_and_markers.R
\name{get_deg_metrics}
\alias{get_deg_metrics}
\title{get DEG metrics from groups}
\usage{
get_deg_metrics(seed, query, params)
}
\arguments{
\item{seed}{count matrix for group 1}

\item{query}{count matrix for group 2}

\item{params}{parameter list (pval,log_fc, min_pct)}
}
\description{
get DEG metrics from groups
}
\details{
Computes basic metrics such as fold change 
and percent of cells containing each gene. This functions 
is used to filter out genes so we don't run differential 
expression on all genes.
importFrom pwr pwr.2p2n.test
}
