% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{get_embeddings}
\alias{get_embeddings}
\title{get embeddings from vesalius_assay}
\usage{
get_embeddings(vesalius_assay, active = TRUE)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay}
}
\value{
embedding matrix
}
\description{
get embeddings from vesalius_assay
}
