% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{get_features}
\alias{get_features}
\title{get cell signal from vesalius assays}
\usage{
get_features(
  seed_assay,
  query_assay,
  signal,
  dimensions = seq(1:30),
  use_norm = "raw",
  verbose = TRUE
)
}
\arguments{
\item{seed_assay}{vesalius_assay object}

\item{query_assay}{vesalius_assay object}

\item{signal}{character string where the signal should be taken from}

\item{dimensions}{int vector - if signal is embeddings which 
embeddings should be selected}

\item{use_norm}{charcater string which counts should be use when
extracting signal}

\item{verbose}{logical - should progress messages be outputed.}
}
\value{
list contain seed signal, query signal and features used
}
\description{
get cell signal from vesalius assays
}
