% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{get_func_from_commit}
\alias{get_func_from_commit}
\title{get function name from commit list}
\usage{
get_func_from_commit(commit)
}
\arguments{
\item{commit}{commit list}
}
\value{
function name
}
\description{
get function name from commit list
}
