% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{get_last_embedding}
\alias{get_last_embedding}
\title{get last embedding}
\usage{
get_last_embedding(vesalius_assay)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}
}
\value{
character string with name of last embedding used
}
\description{
get last embedding
}
