% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{get_mapping_scores}
\alias{get_mapping_scores}
\title{get mapping scores from vesalius assay}
\usage{
get_mapping_scores(vesalius_assay)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}
}
\value{
data.frame containing mapping scores between reference and query cells
}
\description{
get mapping scores from vesalius assay
}
