% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{get_markers}
\alias{get_markers}
\title{get markers from vesalius_assay}
\usage{
get_markers(vesalius_assay, trial = "last")
}
\arguments{
\item{vesalius_assay}{a vesalius_assay}

\item{trial}{character string describing name of DEG trial 
to retrieve.}
}
\value{
marker data frame
}
\description{
get markers from vesalius_assay
}
