% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{get_neighborhood}
\alias{get_neighborhood}
\title{Method dispatch function for neighborhood selection}
\usage{
get_neighborhood(coord, signal, method, k = 20, depth = 3, radius = 20)
}
\arguments{
\item{coord}{data.frame - coordinates of assay (barcodes, x, y)}

\item{signal}{matrix - matrix or sparse matrix containing assay 
signal for all spatial indices contained in coord}

\item{method}{character - which method should be use to collect 
neighborhood - switch matches}

\item{k}{int - how many nearest neighbors from KNN algorithm}

\item{depth}{int - graph path depth to define neighborhood 
0 = self, 1 = direct neigbors, 2 = neighbors of neighbors, etc}

\item{radius}{- numeric - radius around center cell}
}
\value{
matrix of average signals for each spatial index and its 
neighborhood.
}
\description{
Method dispatch function for neighborhood selection
}
