% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{get_ranges}
\alias{get_ranges}
\title{get coordinate ranges}
\usage{
get_ranges(path)
}
\arguments{
\item{path}{list of x and y coordinates for each terrirtory}
}
\value{
list containing ranges in x dim and y dim
}
\description{
get coordinate ranges
}
