% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getSeuratCoordinates}
\alias{getSeuratCoordinates}
\title{getSeuratCoordinates get barcode coordinates from a seurat object}
\usage{
getSeuratCoordinates(seurat)
}
\arguments{
\item{seurat}{Seurat object containing all barcode values from ST assay}
}
\value{
Seurat object only containing the desired barcodes.
}
\description{
getSeuratCoordinates get barcode coordinates from a seurat object
}
\details{
Essentially a wrapper function to the Seurat 
\code{GetTissueCoordinates} function. 
The function serves mainly the purpose of a place holder function for 
future iterations of Vesalius.
}
\examples{
\dontrun{
data(Vesalius)
}
}
