% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{globalise_territories}
\alias{globalise_territories}
\title{globalise territories
converts territories number from territory values per segment
to territory values overall the entire spatial omics assay.}
\usage{
globalise_territories(img)
}
\arguments{
\item{img}{a vesalius territory data frame containing segments
and new territory trial}
}
\value{
the same data frame but with adjusted terriroty values
}
\description{
globalise territories
converts territories number from territory values per segment
to territory values overall the entire spatial omics assay.
}
