% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metrics.R
\name{hclust_scores}
\alias{hclust_scores}
\title{Hiearchal clustering of mapping scores for co mapping events}
\usage{
hclust_scores(score, h, k, verbose)
}
\arguments{
\item{score}{matrix containing mapping scores}

\item{h}{numeric representing the normalized height that will be 
used for clustering (normalized to ensure that it will be between 0 and 1)}

\item{k}{int - number of clusters}

\item{verbose}{logical - print progress messages}
}
\description{
Hiearchal clustering of mapping scores for co mapping events
}
