% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{image_entropy}
\alias{image_entropy}
\title{compute information content of image using entropy}
\usage{
image_entropy(active)
}
\arguments{
\item{active}{data frame containing pixel values as cimg df}
}
\value{
image entropy
}
\description{
compute information content of image using entropy
}
