% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{image_plot}
\alias{image_plot}
\title{imagePlot - Vesalius image plotting}
\usage{
image_plot(
  vesalius_assay,
  dimensions = seq(1, 3),
  embedding = "last",
  cex = 10
)
}
\arguments{
\item{cex}{numeric - font and point size resizing factor.}

\item{image}{a Vesalius data frame containing at least
barcodes, x, y, cc, value}

\item{as.cimg}{logical - If TRUE, will plot directly using cimg plotting.
If FALSE, will produce a ggplot object.}
}
\value{
cimg plot or ggplot object
}
\description{
imagePlot - Vesalius image plotting
}
\details{
Vesalius provides basic plotting functions. This function
only shows Vesalius images (i.e. using true embedded colours or segmented
colours).

Currently, we advise to use as.cimg as false for custom plots.

As Vesalius nearly always returns data frames, custom plots are left to
the discretion of the user.
}
\examples{
\dontrun{
data(vesalius)
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
imagePlot(image)
# as ggplot
g <- imagePlot(image,as.cimg = F)
}
}
