% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{imagePlot}
\alias{imagePlot}
\title{imagePlot - Vesalius image plotting}
\usage{
imagePlot(image, as.cimg = TRUE, cex = 1)
}
\arguments{
\item{image}{a Vesalius data frame containing at least
barcodes, x, y, cc, value}

\item{as.cimg}{logical - If TRUE, will plot directly using cimg plotting.
If FALSE, will produce a ggplot object.}

\item{cex}{numeric - font and point size resizing factor.}
}
\value{
cimg plot or ggplot object
}
\description{
imagePlot - Vesalius image plotting
}
\details{
Vesalius provides basic plotting functions. This function
only shows Vesalius images (i.e. using true embedded colours or segmented
colours).

Currently, we advise to use as.cimg as false for custom plots.

As Vesalius nearly always returns data frames, custom plots are left to
the discretion of the user.
}
\examples{
\dontrun{
data(Vesalius)
}
}
