% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{index}
\alias{index}
\title{calculate weighted index of similarity between 2 signals}
\usage{
index(
  seed_path,
  query_path,
  partials = 500,
  threshold = 0.7,
  weight = rep(0.3, 3)
)
}
\arguments{
\item{seed_path}{gene signal of seed territories}

\item{query_path}{gene signal of query territories}

\item{partials}{int - number of FFT partials to use}

\item{threshold}{numeric time/freq threshold}

\item{weight}{vector of 3 numeric weights - weight used in index calcultion}
}
\value{
matrix of similarity indices between each territory
}
\description{
calculate weighted index of similarity between 2 signals
}
