% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{initialize_matches}
\alias{initialize_matches}
\title{initialize matched data frame}
\usage{
initialize_matches(cost_matrix)
}
\arguments{
\item{cost_matrix}{matrix containing cost of each cell pair}
}
\value{
data.frame templated which will contain the best matching pairs
}
\description{
initialize matched data frame
}
