% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{int_sctransform}
\alias{int_sctransform}
\title{SCTransform}
\usage{
int_sctransform(counts, assay = "Spatial", nfeatures)
}
\arguments{
\item{counts}{seurat object containing counts}

\item{assay}{seurat assay to use here we jjust use the spatial one
but we could just use default RNA}

\item{nfeatures}{number of top variable features to select}
}
\value{
list with seurat object used later and normalised counts to be stored
in a vesalius object
}
\description{
SCTransform pre-processing from Seurat
}
