% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate_assays}
\alias{integrate_assays}
\title{integrate 2 vesalius assays}
\usage{
integrate_assays(
  mapped,
  reference,
  method = "CCAIntegration",
  nfeatures = 2000,
  signal = "variable_features",
  dimensions = 30,
  infer = TRUE,
  use_counts = "raw",
  labels_mapped = NULL,
  labels_reference = NULL,
  verbose = TRUE
)
}
\arguments{
\item{mapped}{vesalius_assay object - mapped veslius assay (map_assays)}

\item{reference}{vesalius_assay object - reference vesalius_assay (seed)}

\item{method}{character - count integration method (methods provided by 
Seurat v5)}

\item{nfeatures}{integer - number of variable features to sue during 
integration.}

\item{signal}{character - defining which signal should be returned:
variable_features, all_features or custom gene list.}

\item{dimensions}{interger - number of dimensions integrated latent space
dimensions.}

\item{infer}{logical - back infer original counts by reversing reduced 
dimensional space roations.}

\item{use_counts}{character - which count matrix to use during integration}

\item{labels_mapped}{character - which columns in the mapped data assay
should be merged with the reference data (see details)}

\item{labels_reference}{character - which columns in the reference data assay
should be merged with the mapped data (see details)}

\item{verbose}{logical - should progressed message be printed}
}
\value{
a vesalius_assay object
}
\description{
integrate 2 vesalius assays
}
\details{
After mapping coordinates from a query onto a reference, vesalius
provides a way to then integrate the assays together. This function will:

* Integrate Counts using Seurat 
* Merge coordinates (adding a jitter to avoid overlapping coordinayes)
* Merge territories (pair-wise merging using labels_mapped and labels_reference 
- everything else will have a separate column).

We also infer log nomalized counts from CCA latent space. 

The final output of this function is a vesalius_assay object containing 
coordinates from the mapped and reference, integrated latent space (e.g.CCA)
integrated counts, merged territories, an additional meta data.

It should be noted that this object does not contain tiles. To use this 
vesalius_assay as any other, add tiles by using \code{\link{generate_tiles}}
}
