% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{integrate_vertically}
\alias{integrate_vertically}
\title{Integrate jointly measured spatial omic assays}
\usage{
integrate_vertically(
  mod1,
  mod2,
  dimensions = seq(1, 30),
  embedding = "last",
  method = "interlace",
  norm_method = "raw",
  dim_reduction = "PCA",
  verbose = TRUE
)
}
\arguments{
\item{mod1}{vesalius_assay object containing first modality}

\item{mod2}{vesalius_assay objecty containing second modality}

\item{dimensions}{numeric vector describing latent space dimensions 
to use during intergration}

\item{method}{character - integration method. interlace - mean - concat 
are available options}

\item{norm_method}{character - which count values should be use 
for integration when using concat method}

\item{dim_reduction}{characater - which dim reduction methods should be 
used for concat integration (PCA,PCA_L,UMAP,LSI,LSI_UMAP,NMF)}

\item{verbose}{logical - should progress message be outputed to the 
console.}
}
\description{
Integrate jointly measured spatial omic assays
}
