% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{interlace_embeds}
\alias{interlace_embeds}
\title{interlace image stack between seed and query}
\usage{
interlace_embeds(seed, query, dimensions)
}
\arguments{
\item{seed}{matrix - seed embedding image stack}

\item{query}{matrix - query embedding image stack}

\item{dimensions}{int vector describing which embeddings
should be selected}
}
\value{
embedding matrix containing seed embeddings + query 
embeddings.
}
\description{
interlace image stack between seed and query
}
\details{
Takes selected embedding from seed and query and 
creates and interlaced embedding matrix starting with the 
first seed embedding
}
