% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_image.R
\name{internal_smooth}
\alias{internal_smooth}
\title{internal smoothing function
Function that does the smoothing on individual image array}
\usage{
internal_smooth(
  image,
  method = c("median", "iso", "box"),
  iter = 1,
  sigma = 1,
  box = 20,
  threshold = 0,
  neuman = TRUE,
  gaussian = TRUE,
  na.rm = FALSE,
  across_levels = "min"
)
}
\arguments{
\item{image}{cimg array}

\item{method}{character describing smoothing method to use "median" ,
"iso"  or "box" or a combination of them.}

\item{iter}{numeric - number of smoothing iteration}

\item{sigma}{numeric - standard deviation associated with isoblur (Gaussian)}

\item{box}{numeric describing box size (centered around center pixel)
for smoothing}

\item{threshold}{numeric - discard pixels that are too low in value (cutoff
threshold only applied in box/median blurs).}

\item{neuman}{logical describing If Neumann boundary conditions should be
used, Dirichlet otherwise (default true, Neumann)}

\item{gaussian}{logical - use gaussian filter}

\item{na.rm}{logical describing if NA values should be removed}

\item{across_levels}{character - method used to account for multiple
smoothing levels (see details). Select from: "min","mean", "max"}
}
\description{
internal smoothing function
Function that does the smoothing on individual image array
}
