% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsolatingTerritories.R
\name{isolateTerritories}
\alias{isolateTerritories}
\title{isolating territories from spatial transcriptomic data}
\usage{
isolateTerritories(img, colDepth = 8, dropRadius = 0.025)
}
\arguments{
\item{img}{data frame with barcodes, xcoord, ycoord, R, G, and B as coloumns}

\item{colDepth}{integer describing the number of colours to extract}

\item{dropRadius}{numeric describing the proportion of total distance to consider when pooling beads together}
}
\value{
a data frame with barcodes, xcoord, ycoord, R, G, B, cluster number, territory number within that cluster.
}
\description{
isolating territories from spatial transcriptomic data
}
\details{
In order to select territories, colours will be collapsed into \code{colDepth} number of
dominant colours. For each colour, territories will be isolated based on distance between point.
The \code{dropRadius} represents the proportion of total distance (x and y values may differ) to use in order
to pool beads/spots together. The algorithm selects a point and pools all neighboring beads into a territory.
The process is applied to the nearest neighbors until no more points can be pooled into the territory.
If beads remain for a given colour, the process is repeated until all beads/spots are put into a territory.
}
