% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jitter_counts}
\alias{jitter_counts}
\title{Count matrix for vesalius - alt}
\format{
sparse count matrix
\describe{
 \item{counts}{Sampled and jittered counts from Puck_200815 taken from slide-seV2}
}
}
\source{
\url{https://singlecell.broadinstitute.org/single_cell/study/SCP948/robust-decomposition-of-cell-type-mixtures-in-spatial-transcriptomics#study-download}
}
\usage{
data(vesalius)
}
\description{
Count matrix for vesalius - alt
}
\keyword{datasets}
