% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{joint_territories}
\alias{joint_territories}
\title{Jointly measured spatial omic assays territories}
\usage{
joint_territories(
  mod1,
  mod2,
  dimensions = seq(1, 30),
  embedding = "last",
  method = "interlace",
  norm_method = "log_norm",
  dim_reduction = "PCA",
  signal = "variable_features",
  verbose = TRUE
)
}
\arguments{
\item{mod1}{vesalius_assay object containing first modality}

\item{mod2}{vesalius_assay objecty containing second modality}

\item{dimensions}{numeric vector describing latent space dimensions 
to use during intergration}

\item{method}{character - integration method. interlace - mean - concat 
are available options}

\item{norm_method}{character - which count values should be use 
for integration when using concat method}

\item{dim_reduction}{characater - which dim reduction methods should be 
used for concat integration (PCA,PCA_L,UMAP,LSI,LSI_UMAP,NMF)}

\item{verbose}{logical - should progress message be outputed to the 
console.}
}
\value{
vesalius object containing new image embeddings
}
\description{
Jointly measured spatial omic assays territories
}
