% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{kmeans_segmentation}
\alias{kmeans_segmentation}
\title{kmeans segmentation function}
\usage{
kmeans_segmentation(
  vesalius_assay,
  dimensions = seq(1, 3),
  col_depth = 10,
  embedding = "last",
  use_center = TRUE,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}

\item{col_depth}{integer or vector of positive integers.
Colour depth used for segmentation. (see details)}

\item{use_center}{logical - If TRUE, only the center pixel value will be used
during segmentation. If FALSE, all pixels will be used (see details)}

\item{cores}{numeric - number of cores that should be used}

\item{verbose}{logical - progress message output.}

\item{vesalius}{a vesalius_assay}

\item{colDepth}{integer or vector of positive integers.
Colour depth used for segmentation.}
}
\value{
list containing segmented image as an active embedding and
territory cluster for all barcodes.
}
\description{
kmeans segmentation function
}
\details{
Run an interaive kmeans segmentation with the possibility 
to run multiple rounds of smoothing.
}
