% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{knn_neighborhood}
\alias{knn_neighborhood}
\title{k nearest neighbors - niche selection}
\usage{
knn_neighborhood(coord, k)
}
\arguments{
\item{coord}{data.frame - coordinates of spatial indices in assay}

\item{k}{int - number of nearest neighbors to select}
}
\value{
list containing barcodes of nearest neighbors for each 
spatial index.
}
\description{
k nearest neighbors - niche selection
}
