% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology_operators.R
\name{layer_territory}
\alias{layer_territory}
\title{extractIdentity compute differentially expressed genes for each territory}
\usage{
layer_territory(
  vesalius,
  territory = NULL,
  trial = "last",
  layer_depth = NULL,
  morphology_factor = 0,
  verbose = RUE
)
}
\arguments{
\item{verbose}{logical - progress message output.}

\item{image}{data.frame - Vesalius formatted data.frame (i.e. barcodes,
x,y,cc,value,cluster, territory)}

\item{seedTerritory}{integer describing which territory should be layered}

\item{layerDepth}{integer describing maximum number of layers}

\item{morphologyFactor}{integer or vector of integers describing growth
and/or shrink extent.}
}
\value{
Returns a Vesalius data.frame with "barcodes","x","y","cc",
"value","tile","cluster","territory" and "layer".
Layer describes the layer to which a barcode belongs.
}
\description{
extractIdentity compute differentially expressed genes for each territory
}
\details{
To divide a territory into layers, one approach is to use edge
detection. Terriotries are converted to a black and white image containing
only the territories of interest. Sobel edge detection is applied along the
x and y axis and all barcodes sharing a pixel with the edge are pooled into
a layer. The process is applied until no more barcodes can be pooled into
a layer.

Morphological operators are applied to the isolated territory prior to
layering (see \code{territoryMorphing}).
}
\examples{
\dontrun{
data("vesalius")
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
# One segmentation round
image <- iterativeSegmentation.array(image)
image <- isolateTerritories.array(image, minBar = 5)
layer <- layerTerritory.edge(image, seedTerritory = 1)
}
}
