% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{layerTerritory.edge}
\alias{layerTerritory.edge}
\title{extractIdentity compute differentially expressed genes for each territory}
\usage{
layerTerritory.edge(
  image,
  seedTerritory = NULL,
  layerDepth = NULL,
  morphologyFactor = 3,
  verbose = TRUE
)
}
\arguments{
\item{image}{data.frame - Vesalius formatted data.frame (i.e. barcodes,
x,y,cc,value,cluster, territory)}

\item{seedTerritory}{integer describing which territory should be layered}

\item{layerDepth}{integer describing maximum number of layers}

\item{morphologyFactor}{integer or vector of integers describing growth 
and/or shrink extent.}

\item{verbose}{logical - progress message output.}
}
\value{
Returns a Vesalius data.frame with "barcodes","x","y","cc",
"value","tile","cluster","territory" and "layer".
Layer describes the layer to which a barcode belongs.
}
\description{
extractIdentity compute differentially expressed genes for each territory
}
\details{
To divide a territory into layers, one approach is to use edge 
detection. Terriotries are converted to a black and white image containing 
only the territories of interest. Sobel edge detection is applied along the 
x and y axis and all barcodes sharing a pixel with the edge are pooled into 
a layer. The process is applied until no more barcodes can be pooled into
a layer. 

Morphological operators are applied to the isolated territory prior to 
layering (see \code{territoryMorphing}).
}
\examples{
\dontrun{
data(Vesalius)
}
}
