% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metrics.R
\name{leiden_scores}
\alias{leiden_scores}
\title{Using Leiden community clustering to for co-mapping events}
\usage{
leiden_scores(score, resolution, nn, verbose)
}
\arguments{
\item{score}{matrix containing mapping scores}

\item{resolution}{numeric - clustering resolution}

\item{nn}{int - number of nearest neighbors to use for graph construction}

\item{verbose}{logical - print output messages}
}
\description{
Using Leiden community clustering to for co-mapping events
}
