% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{log_norm}
\alias{log_norm}
\title{log norm}
\usage{
log_norm(counts, nfeatures)
}
\arguments{
\item{counts}{seurat object containing counts}

\item{nfeatures}{number of top variable features to select}
}
\value{
list with seurat object used later and normalised counts to be stored
in a vesalius object
}
\description{
log normalisation, scaling and top variable features
}
