% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{louvain_segmentation}
\alias{louvain_segmentation}
\title{leiden segmentation}
\usage{
louvain_segmentation(
  vesalius_assay,
  dimensions = seq(1, 3),
  col_resolution = 0.01,
  embedding = "last",
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}

\item{dimensions}{embedding dimensions used for clustering}

\item{col_resolution}{clustering resolution used for leiden}

\item{embedding}{embedding type used for clustering}

\item{cores}{numeric describing the number of cores that should be used}

\item{verbose}{logical if progress message should outputed}
}
\value{
list with updated segmented embedding values 
and segment territories.
}
\description{
using leiden clustering to cluster colors
}
