% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{make_composition_matrix}
\alias{make_composition_matrix}
\title{create a composition matrix from cell type labels in niches}
\usage{
make_composition_matrix(niches)
}
\arguments{
\item{niches}{list of cell labels for each niche in a data set}
}
\value{
a matrix with rows as cell labels and columns as each cell
}
\description{
create a composition matrix from cell type labels in niches
}
\details{
The aim is to reformat the niches to make the scoring faster
}
