% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{match_cells}
\alias{match_cells}
\title{Internal function returning cell type label match as numiric}
\usage{
match_cells(idx, query_labels, seed_labels)
}
\arguments{
\item{idx}{numeric - used to iterate down the labels}

\item{query_labels}{cell type labels for query}

\item{seed_labels}{cell type labels for seed}
}
\value{
binary numeric vector
}
\description{
Internal function returning cell type label match as numiric
}
