% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{match_index}
\alias{match_index}
\title{Finding optimal mapping between query and seed}
\usage{
match_index(cost_matrix)
}
\arguments{
\item{cost_matrix}{matrix with cost between query (rows) and 
seed (columns)}
}
\value{
data.frame with point in query (from) mapped to 
which point in seed (to) as well as cost of mapping for that pair
}
\description{
Finding optimal mapping between query and seed
}
\details{
Uses a Hungarian solver to minimize overall cost of 
pair assignement.
}
