% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{merge_territories}
\alias{merge_territories}
\title{Merge territories using specified labels}
\usage{
merge_territories(
  matched,
  reference,
  coordinates,
  labels_mapped,
  labels_reference
)
}
\arguments{
\item{matched}{matched (query) vesalius_assay object}

\item{reference}{reference vesalius_assay}

\item{coordinates}{coord data frame after merging (see merge_coordinates)}

\item{labels_mapped}{Territory columns to merge in mapped object}

\item{labels_reference}{Territory columns to merge in reference object}
}
\value{
territory data frame with merged columns
}
\description{
Merge territories using specified labels
}
\details{
Territory data frame generally contain a lot of information 
depending on how much analysis the user has run. 
(Segments, Cells, territories, morphology). 
Using pariwise matching of labels parse in labels_mapped and labels_reference
the data will be merged into a single column. 
A new column will be added to distinguish samples. All other columns that are not
present in the labels argument will retain their name and NA will be added 
to cell present in the other data set (similar to left and right joins).
}
