% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{neighbor_expression}
\alias{neighbor_expression}
\title{compute average expression of local neighborhood}
\usage{
neighbor_expression(distance, features)
}
\arguments{
\item{distance}{knn neighbor matrix}

\item{features}{count matrix or feature matrix to average}
}
\value{
average feature matrix. The expression of each cell 
is replace by the average expression of the k nearest neighbors
}
\description{
compute average expression of local neighborhood
}
