% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{neighborhood_signal}
\alias{neighborhood_signal}
\title{compute average expression of local neighborhood}
\usage{
neighborhood_signal(neighbors, signal)
}
\arguments{
\item{neighbors}{list of local neighbors}

\item{signal}{count matrix or feature matrix to average}
}
\value{
average feature matrix. The expression of each cell 
is replace by the average expression of the k nearest neighbors
}
\description{
compute average expression of local neighborhood
}
