% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{no_norm}
\alias{no_norm}
\title{no norm}
\usage{
no_norm(counts, use_count = "raw")
}
\arguments{
\item{counts}{seurat object containing counts}

\item{use_count}{string describing name that needs to be added to
list element. This list will be appended to the count slot in
the vesalius_assay.}
}
\value{
list with seurat object used later and raw counts to be stored in
the vesalius objects
}
\description{
no normalisation applied simply return raw counts
}
\details{
Here, either the user doesn't want to normalise the data or
they provide their custom count matrix. In this case, we parse it 
as "none" to avoid writing another function and add the custom name.
}
