% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{normalize_path}
\alias{normalize_path}
\title{normalise coordinates of territory edge}
\usage{
normalize_path(path, global = TRUE)
}
\arguments{
\item{path}{data frame of coordinates corresponding to 
the x and y coordinates of the edge of each territory.}

\item{global}{logical - using global mion max values or local.}
}
\value{
list of nromalised x and y coordinates
}
\description{
normalise coordinates of territory edge
}
