% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{optimize_matching}
\alias{optimize_matching}
\title{Wrapper for fast niche composition}
\usage{
optimize_matching(
  cost_matrix,
  batch_size = 10000,
  epochs = 1,
  allow_duplicates = TRUE,
  verbose = TRUE
)
}
\description{
Wrapper for fast niche composition
}
