% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metrics.R
\name{overlap_distance_matrix}
\alias{overlap_distance_matrix}
\title{create a distance matrix based on mapping scores overlaps}
\usage{
overlap_distance_matrix(score, top_nn, verbose)
}
\arguments{
\item{score}{matrix - cost matrix}

\item{top_nn}{integer nearest neighbors}

\item{dec}{logical - if order should be decreasing i.e return cost or score}
}
\description{
create a distance matrix based on mapping scores overlaps
}
