% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{pearson_approx}
\alias{pearson_approx}
\title{compute fast pearson correlation between matrices}
\usage{
pearson_approx(seed, query)
}
\arguments{
\item{seed}{(r x c) seed matrix}

\item{query}{(r x c') query matrix}
}
\value{
correlation matrix between each columns (c' x c)
}
\description{
compute fast pearson correlation between matrices
}
