% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_image.R
\name{populate_graph}
\alias{populate_graph}
\title{populate graph with network connections}
\usage{
populate_graph(chunk)
}
\arguments{
\item{chunk}{data frame with nearest neighbors}
}
\value{
data frame with network connections
}
\description{
populate graph with network connections
}
