% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{process_counts}
\alias{process_counts}
\title{process counts}
\usage{
process_counts(
  counts,
  assay,
  method = "log_norm",
  nfeatures = 2000,
  min_cutoff = "q5",
  verbose = TRUE
)
}
\arguments{
\item{counts}{count matrix in the form of a sparse matrix}

\item{assay}{character string describing the assay that is being pre-processed
in the vesaliusObject or vesalius_assay}

\item{method}{character string describing which normalisation method to use.
One of the following "log_norm", "SCT", "TFIDF", "raw".}

\item{nfeatures}{numeric describing the number of variable features to use.}

\item{min_cutoff}{only used when dimensionality reduction method is
LSI or LSI_UMAP cutoff for feature to be included in the 
VariableFeatures for the object.}

\item{verbose}{logical - progress messages outputed or not}
}
\description{
pre-process count matrices
}
