% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{quantile_norm}
\alias{quantile_norm}
\title{quantile normalization}
\usage{
quantile_norm(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
quantile normalised vector
}
\description{
quantile normalization
}
