% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{radius_neighborhood}
\alias{radius_neighborhood}
\title{Radius based method to select niche}
\usage{
radius_neighborhood(coord, radius)
}
\arguments{
\item{coord}{data.frame - coordinates of spatial indices in assay}

\item{radius}{- numeric - radius around center cell}
}
\value{
list containing barcodes of nearest neighbors for each 
spatial index.
}
\description{
Radius based method to select niche
}
