% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{rasterise}
\alias{rasterise}
\title{rasterise tiles}
\usage{
rasterise(filtered)
}
\arguments{
\item{filtered}{data.frame with voronoi tile coordinates}
}
\value{
a data frame barcodes and their associated pixels.
}
\description{
fill tiles with pixel - rasterisation
}
\details{
Here, we take our tile cooridnates and fill them 
with pixels. Essentially, each voronoi tile can be discretised 
into a series of pixels and we achieve this by reconstructing a 
polygon from the tesselation coordinates and finding all discrete 
value that fall within this polygon. 

Note that the polygon coordinates need to be "convexified". 
Essentially, the order of the coordinates maters here so we
order the coordinates before recontructing the polygon (see
\code{\link{convexify}})
}
