% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{rebalence_colors}
\alias{rebalence_colors}
\title{rebalence colors}
\usage{
rebalence_colors(coordinates, dimensions, method = "minmax")
}
\arguments{
\item{coordinates}{data frame containing barcodes, x/y coord, origin,
and color value.}

\item{method}{character string: min max or truncate}

\item{dims}{number dimensions select for plotting}
}
\value{
data frame with [0,1] bound values
}
\description{
rebalence colors
}
\details{
This function is use to re-bound values between 0 and 1. 
Some image processing steps may lead to negative values being introduced.
Imager handles this without an issue but for plotting we want to make sure 
that all values are indeed bound between 0 and 1. We will try to different 
methods either we truncate the value or we min max norm.
}
