% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{reduce_tensor_resolution}
\alias{reduce_tensor_resolution}
\title{reduce tensor resoltuon}
\usage{
reduce_tensor_resolution(coordinates, tensor_resolution = 1)
}
\arguments{
\item{coordinates}{data frame with barcode coordinates}

\item{tensor_resolution}{numeric (range 0 - 1) describing the compression
ratio to be applied to the final image. Default = 1}
}
\value{
a data frame with barcodes, x and coordinates
}
\description{
reduce the size of the image tensor by merging barcodes together after
compressing coordinates values
}
\details{
While we compress the coordinates, we retain all barcodes. 
Each barcode that overlap with each other are marged together. Their 
respective counts will also be merged together. This allows us to 
retain all barcodes for downstream analysis. 


TO DO: replace the tensor resolution reduction with super pixels
}
