% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{register_image}
\alias{register_image}
\title{register_image
register microscopy image to vesalius image}
\usage{
register_image(
  vesalius_assay,
  source = "Territory",
  dimensions = seq(1, 3),
  verbose
)
}
\arguments{
\item{vesalius_assay}{a vesalius_Assay object}

\item{source}{character string - name of source image}

\item{dimensions}{numeric integer - number of dimension that should
be used if either source is an embedding.}

\item{verbose}{logical - should progressed message be rpint to the
console.}
}
\value{
a vesalius_assay object
}
\description{
register_image
register microscopy image to vesalius image
}
