% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{regularise}
\alias{regularise}
\title{Internal regularise function.
performs total variance regularisation}
\usage{
regularise(img, lambda = 1, niter = 100)
}
\arguments{
\item{img}{image array - generally a matrix}

\item{lambda}{lambda value for regularisation}

\item{niter}{numeric number of rounds of regularisation}
}
\description{
Internal regularise function.
performs total variance regularisation
}
