% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{regularise_image}
\alias{regularise_image}
\title{regulariseImage denoise Vesalius images via variance regularization}
\usage{
regularise_image(
  vesalius,
  dimensions = seq(1, 3),
  embedding = "last",
  lambda = 1,
  niter = 100,
  method = "TVL2.FiniteDifference",
  normalise = TRUE,
  na.rm = TRUE,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{lambda}{numeric - positive real numbers describing regularization
parameter (see details)}

\item{niter}{numeric - number of variance regularization iterations
(Default = 100)}

\item{method}{character - regularization method.
Select from: "TVL2.FiniteDifference","TVL1.PrimalDual",and "TVL2.PrimalDual"
(see details)}

\item{normalise}{logical - If TRUE, regularized colour values will be
min max normalised.}

\item{na.rm}{logical - If TRUE, NAs are removed}

\item{verbose}{logical - progress message output.}

\item{image}{data.frame - Vesalius formatted data.frame (i.e. barcodes,
x,y,cc,value,tile,...)}

\item{invert}{logical - If TRUE, colours will be inverted i.e. 1 - colorValue
(background set to 1 instead of 0).}
}
\value{
Returns a Vesalius data.frame with "barcodes","x","y","cc",
"value","tile",etc
}
\description{
regulariseImage denoise Vesalius images via variance regularization
}
\details{
Image regularization can be seen as a form of image denoising.
Details on each method can be found in the tvR package under the denoise2
function \href{tvR}{https://cran.r-project.org/web/packages/tvR/tvR.pdf}.

We recommend using the TVL2.FiniteDifference method.
}
\examples{
\dontrun{
data(vesalius)
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
image <- regulariseImage(image)
}
}
