% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{rename_counts}
\alias{rename_counts}
\title{renaming counts to remain consistent with vesalius nomenclature}
\usage{
rename_counts(
  counts,
  seed_cells,
  seed_genes,
  query_cells,
  query_genes,
  features
)
}
\description{
renaming counts to remain consistent with vesalius nomenclature
}
