% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_to_RGB.R
\name{rgbUMAP}
\alias{rgbUMAP}
\title{Embed 3D UMAP projections into RGB colour space}
\usage{
rgbUMAP(SO, pcs = 30, adjusted = FALSE, conserveSparse = TRUE, verbose = TRUE)
}
\arguments{
\item{SO}{Seurat Object after normalisation, scaling and
finding variable features.}

\item{pcs}{number of Principle components to use for UMAP projections}

\item{adjusted}{logical indicating if RGB code should be adjusted to
colour contribution per channel.}

\item{conserveSparse}{logical indicating if sparse matrix format
should be conserved.}

\item{verbose}{logical - progress message output.}
}
\value{
Returns a Vesalius data.frame with "barcodes","x","y","cc",
"value","slice"
}
\description{
Embed 3D UMAP projections into RGB colour space
}
\details{
Emedding 3D UMAP projections into the RGB color space is quite
straight forward: Each UMAP dimension is extracted and min/max normalised to
ensure that each value reanges between 0 and 1. The first dimension is
assigned to red color channel, the second dimension to the green color
channel and the third dimension assgined to the blue color channel.
UMAP projections consider more varience per channel than PCA embeddings but
this additional information in further compressed. Using PCA slice provides
a way to visualize and explore different anatomical structure in more depth.
Choosing between rgbPCA and rgbUMAP will depend on the question at hand and
the depth of the analysis.
}
\examples{
\dontrun{
data(Vesalius)
}
}
