% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{score_graph}
\alias{score_graph}
\title{importFrom future.apply future_lapply}
\usage{
score_graph(graph, signal, scoring_method = "spearman", verbose = TRUE)
}
\description{
importFrom future.apply future_lapply
}
