% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{segment_gradient}
\alias{segment_gradient}
\title{Segment gradients from images}
\usage{
segment_gradient(
  vesalius_assay,
  dimensions = seq(1, 3),
  embedding = "last",
  gradients = 10,
  kernel = NULL,
  verbose = TRUE
)
}
\description{
Segment gradients from images
}
