% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{select_dimensions}
\alias{select_dimensions}
\title{select latent space dimensions}
\usage{
select_dimensions(
  vesalius_assay,
  method = "entropy",
  threshold = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay object}

\item{method}{character string describing which method should be use to 
select latent space dimensions}

\item{threshold}{numeric - information content threshold}
}
\value{
a numeric vector of suggested dimensions to use for analysis.
}
\description{
select latent space dimensions
}
\details{
Since each latent space dimension is represented as an image, 
we can compute the information content of each image. If an image drops below
a certain threhsold, we assume that it mainly contains noise and will not
be uesful for territtory selection.

Currently, we compute the information content using entropy.

Vesalius will always compute information content on the active embedding.
Since the active embedding can be subjected to image processing steps,
we recommend selecting dimensions after image processing and prior to 
image segmentation.
}
\examples{
\dontrun{
data(vesalius)
# First we build a simple object
ves <- build_vesalius_object(coordinates, counts)
# We can do a simple run 
ves <- generate_embeddings(ves)
ves <- smooth_image(ves, iter = 5, sigma = 3)
embeds <- select_dimensions(ves)

}
}
