% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_pixel.R
\name{select_initial_indices}
\alias{select_initial_indices}
\title{select inital indices}
\usage{
select_initial_indices(coordinates, k = 500)
}
\arguments{
\item{coordinates}{data frame containing spatial coordinates}

\item{k}{numeric - number of intial coordinates}
}
\value{
barcodes of starting coordinates
}
\description{
select inital indices
}
