% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{signal_similarity}
\alias{signal_similarity}
\title{compute the similarity between seed and query signals}
\usage{
signal_similarity(seed, query, method = "pearson", digits = 4)
}
\arguments{
\item{seed}{seed signal}

\item{query}{query signal}

\item{method}{character - which pearson correlation method to use}

\item{digits}{int - number of digits to retain after computing 
correlations. Less digits = faster mapping}
}
\value{
matrix with query as rows and seed as colmuns
}
\description{
compute the similarity between seed and query signals
}
\details{
Chunking cost and signal into smaller chunks to run the 
correlation score in paralell. There is room for improvement here.
Tested a few ways to computed pearson correlations for speed.
Pearson exact is the slowest but pearson and pearson_fast give similar
results. Can be changed for other metrics later. Cosine distance?
}
