% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{signal_similiarity}
\alias{signal_similiarity}
\title{compute signal similarity between 2 sets of territories}
\usage{
signal_similiarity(
  seed_path,
  query_path,
  domain = "index",
  partials = 200,
  threshold = 0.7,
  weight = rep(0.33, 3)
)
}
\arguments{
\item{seed_path}{list of x and y coordinates of seed territories}

\item{query_path}{list of x and y coordinates of query territories}

\item{domain}{similarity method to use (index or dtw)}

\item{partials}{int - number of FFT partials to use}

\item{threshold}{numeric time/freq threshold}

\item{weight}{vector of 3 numeric weights - weight used in index calcultion}
}
\value{
a matrix containing similarity scores between each territory in each 
data set.
}
\description{
compute signal similarity between 2 sets of territories
}
