% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_messages.R
\name{simple_bar}
\alias{simple_bar}
\title{Simple progress bar output}
\usage{
simple_bar(verbose = TRUE)
}
\arguments{
\item{verbose}{logical - whether to print output}
}
\description{
Simple progress bar output
}
