% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spix.R
\name{slic_segmentation}
\alias{slic_segmentation}
\title{SLIC segmentation}
\usage{
slic_segmentation(
  vesalius_assay,
  dimensions,
  col_resolution,
  embedding,
  index_selection = "bubble",
  compactness = 1,
  scaling = 0.3,
  threshold = 0.9,
  max_iter = 1000,
  verbose
)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}

\item{dimensions}{dimensions to use}

\item{col_resolution}{color resolution for segmentation}

\item{embedding}{embedding to use}

\item{index_selection}{method for selecting initial indices}

\item{compactness}{compactness factor}

\item{scaling}{scaling factor}

\item{threshold}{correlation threshold}

\item{max_iter}{int - max number of kmeans iterations}

\item{verbose}{logical - progress messages}
}
\value{
segmented territories
}
\description{
SLIC segmentation
}
