% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{slot_apply}
\alias{slot_apply}
\title{apply function to slots in object}
\usage{
slot_apply(x, func, ...)
}
\arguments{
\item{x}{S4 class object}

\item{func}{closure - function to be applied to slot}

\item{...}{any other argument to parse to func}
}
\description{
apply function to slots in object
}
