% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{slot_assign}
\alias{slot_assign}
\title{assing value to slot in object if nested in list}
\usage{
slot_assign(object, sl, input)
}
\arguments{
\item{object}{list containing S4 class objects}

\item{sl}{string - slot name}

\item{input}{data that should be assign to slot in S4 object}
}
\description{
assing value to slot in object if nested in list
}
