% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_utilities_and_methods.R
\name{slot_get}
\alias{slot_get}
\title{get specific slot out of object if nested in list}
\usage{
slot_get(object, slot_name)
}
\arguments{
\item{object}{list containing S4 class objects}

\item{slot_name}{string name of slot to extract}
}
\description{
get specific slot out of object if nested in list
}
