% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsolatingTerritories.R
\name{smoothArray}
\alias{smoothArray}
\title{Smoothing image array - wrapper function for imager smoothing functions}
\usage{
smoothArray(
  img,
  method = c("median", "iso", "box"),
  sigma = 1,
  box = 20,
  threshold = 0,
  neuman = TRUE,
  gaussian = TRUE,
  na.rm = FALSE,
  iter = 1
)
}
\arguments{
\item{img}{4 dimensional array or cimg object}

\item{method}{character describing smoothing method to use "median" , "iso"  or "box"}

\item{sigma}{numeric - standard deviation associated with the isoblur}

\item{box}{numeric describing box size (centered around center pixel) for smoothing}

\item{threshold}{numeric - discard pixels that are too low in value (see \code{medianblur})}

\item{neuman}{logical describing If Neumann boundary conditions should be used,
Dirichlet otherwise (default true, Neumann)}

\item{gaussian}{logical - use gaussian filter}

\item{na.rm}{logical describing if NA values should be removed}

\item{iter}{numeric describing number of smoothing rounds to apply to image}
}
\value{
Array or cimg object after bieng smoothed
}
\description{
Smoothing image array - wrapper function for imager smoothing functions
}
\details{
Applying smoothing to image data.
}
