% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsolatingTerritories.R
\name{smoothToDominant}
\alias{smoothToDominant}
\title{Applying morphological filter to beads by smoothing bead to dominant colour of nearest neighbors}
\usage{
smoothToDominant(img, nn, k = 10, threshold = 0.5)
}
\arguments{
\item{img}{data frame with barcodes, xcoord, ycoord, R, G, and B as coloumns}

\item{nn}{nearest neighbors list}

\item{k}{numeric describing number of neighbors to use for smoothing}

\item{threshold}{numeric between 0 and 1 describing which quantile should be used to select dominant colour.}
}
\value{
data frame with barcodes, xcoord, ycoord, R, G, and B as coloumns but with adjusted RGB values.
}
\description{
Applying morphological filter to beads by smoothing bead to dominant colour of nearest neighbors
}
\details{
Applying morphological filters on non image data. Morphological filters converts pixels values
to the average value of neighboring pixels. The same process is applied here however instead of using
pixel, this function uses closest beads/spots in space (according to x/coordinates). Depending on how nearest
neighbors were selected, nearest neighbors should be beads/spots that surround the center bead in all directions.
}
