% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{som_map}
\alias{som_map}
\title{importFrom kohonen som map scale somgrid}
\usage{
som_map(seed_trial, seed_signal, query_trial, query_signal, anchors)
}
\description{
importFrom kohonen som map scale somgrid
}
