% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subSetTerritories}
\alias{subSetTerritories}
\title{subSetTerritories gets all barcodes associated with a territory from a
Seurat object.}
\usage{
subSetTerritories(territories, seurat)
}
\arguments{
\item{territories}{vector of barcode values (generally character strings)}

\item{seurat}{Seurat object containing all barcode values from ST assay}
}
\value{
Seurat object only containing the desired barcodes.
}
\description{
subSetTerritories gets all barcodes associated with a territory from a
Seurat object.
}
\details{
Essentially a wrapper function to the Seurat \code{subset} function.
The function serves mainly the purpose of a place holder function for
future iterations of Vesalius.
}
\examples{
\dontrun{
data(vesalius)
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
# One segmentation round
image <- iterativeSegmentation.array(image)
image <- isolateTerritories.array(image, minBar = 5)
# subset 
sub <- subSetTerritories(image$barcodes[image$territory ==1],vesalius)

}
}
