% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{territory_neighborhood}
\alias{territory_neighborhood}
\title{Territory selection for large scale niche}
\usage{
territory_neighborhood(coord)
}
\arguments{
\item{coord}{data.frame - coordinates of spatial indices in assay}
}
\value{
list containing barcodes of spatial indices in each territory
}
\description{
Territory selection for large scale niche
}
