% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{territory_plot}
\alias{territory_plot}
\title{territoryPlot plotting Vesalius territories}
\usage{
territory_plot(
  vesalius_assay,
  trial = "last",
  split = FALSE,
  randomise = TRUE,
  cex = 10,
  cex_pt = 1
)
}
\arguments{
\item{split}{logical - If TRUE, territories will be plotted in
separate panels}

\item{randomise}{logical - If TRUE, colour palette will be randomised.}

\item{cex}{numeric describing font size multiplier.}

\item{territories}{a Vesalius data frame conatining territory information
(i.e. containing the terriotry column)}

\item{cex.pt}{numeric describing point size multiplier.}
}
\value{
a ggplot object
}
\description{
territoryPlot plotting Vesalius territories
}
\details{
Territory plots show all territories in false colour after they
have been isolated from a Vesalius image.
}
\examples{
\dontrun{
data(vesalius)
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
# One segmentation round
image <- iterativeSegmentation.array(image)
image <- isolateTerritories.array(image, minBar = 5)
g <- territoryPlot(image,cex = 12, cex.pt = 1)
}
}
