% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{territoryMorphing}
\alias{territoryMorphing}
\title{territoryMorphing applies morphological operators to a set of territoriees}
\usage{
territoryMorphing(territories, morphologyFactor, image, verbose = TRUE)
}
\arguments{
\item{territories}{data.frame - Vesalius formatted data.frame (i.e. barcodes,
x,y,cc,value,cluster, territory) only containing territories of interest.}

\item{morphologyFactor}{integer or vector of integers describing growth
and/or shrink extent.}

\item{image}{data.frame - Vesalius formatted data.frame (i.e. barcodes,
x,y,cc,value,cluster, territory) containing all territories.}

\item{verbose}{logical - progress message output.}
}
\value{
Returns a Vesalius data.frame with "barcodes","x","y","cc",
"value","tile","cluster", and "territory".
This data.frame will contain the new territory after morphological operators
and will contain barcodes associated with other territories and colour
segments.
}
\description{
territoryMorphing applies morphological operators to a set of territoriees
}
\details{
Territory morphing can manipulate territories by growing, shrinking,
filling, and cleaning territories.
Growing = Positive integers - Territory will be dilated by x number of pixels
Shrinking = Negative integers - Territory will be contracted by x number of
pixels
Filling = grow followed by shrink.
Cleaning = shrink followed by grow.
}
\examples{
\dontrun{
data(Vesalius)
}
}
