% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{territoryPlot}
\alias{territoryPlot}
\title{territoryPlot plotting Vesalius territories}
\usage{
territoryPlot(
  territories,
  split = FALSE,
  randomise = TRUE,
  cex = 1,
  cex.pt = 0.25
)
}
\arguments{
\item{territories}{a Vesalius data frame conatining territory information
(i.e. containing the terriotry column)}

\item{split}{logical - If TRUE, territories will be plotted in
separate panels}

\item{randomise}{logical - If TRUE, colour palette will be randomised.}

\item{cex}{numeric describing font size multiplier.}

\item{cex.pt}{numeric describing point size multiplier.}
}
\value{
a ggplot object
}
\description{
territoryPlot plotting Vesalius territories
}
\details{
Territory plots show all territories in false colour after they
have been isolated from a Vesalius image.
}
\examples{
\dontrun{
data(Vesalius)
}
}
