% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{top_cluster}
\alias{top_cluster}
\title{top cluster
get cluster that is most represented in tile}
\usage{
top_cluster(cluster)
}
\arguments{
\item{cluster}{named vector containing cluster representation for a barcode}
}
\description{
top cluster
get cluster that is most represented in tile
}
