% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{ves_altieri}
\alias{ves_altieri}
\title{compute Altieri spatial entropy}
\usage{
ves_altieri(active)
}
\arguments{
\item{active}{data frame containing pixel values as cimg df}
}
\value{
image entropy
}
\description{
compute Altieri spatial entropy
}
