% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{ves_batty}
\alias{ves_batty}
\title{compute batty spatial entropy}
\usage{
ves_batty(active)
}
\arguments{
\item{active}{data frame containing pixel values as cimg df}
}
\value{
image entropy
}
\description{
compute batty spatial entropy
}
