% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{ves_contagion}
\alias{ves_contagion}
\title{compute relative contagion index}
\usage{
ves_contagion(active)
}
\arguments{
\item{active}{data frame containing pixel values as cimg df}
}
\value{
contagion index
}
\description{
compute relative contagion index
}
