% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{ves_karlstrom}
\alias{ves_karlstrom}
\title{compute karlstrom entropy}
\usage{
ves_karlstrom(active)
}
\arguments{
\item{active}{data frame containing pixel values as cimg df}
}
\value{
entropy
}
\description{
compute karlstrom entropy
}
