% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{ves_leibovici}
\alias{ves_leibovici}
\title{compute leibovici entropy}
\usage{
ves_leibovici(active)
}
\arguments{
\item{active}{data frame containing pixel values as cimg df}
}
\value{
entropy
}
\description{
compute leibovici entropy
}
