% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{ves_oneil}
\alias{ves_oneil}
\title{compute oneil entropy}
\usage{
ves_oneil(active)
}
\arguments{
\item{active}{data frame containing pixel values as cimg df}
}
\value{
image entropy
}
\description{
compute oneil entropy
}
