% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{ves_shannon}
\alias{ves_shannon}
\title{compute shannow entropy}
\usage{
ves_shannon(active)
}
\arguments{
\item{active}{data frame containing pixel values as cimg df}
}
\value{
image entropy
}
\description{
compute shannow entropy
}
