% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VesaliusObjects.R
\docType{class}
\name{Vesalius-class}
\alias{Vesalius-class}
\title{Vesalius Object}
\description{
Vesalius objects are place holders for data required to run Vesalius.
}
\section{Slots}{

\describe{
\item{\code{countMatrix}}{matrix containing gene (rows) counts and barcodes (columns)}

\item{\code{cooridinates}}{data.frame containing x/y cooridinates for each barcode (barcode , xcoord,ycoord)}

\item{\code{SO}}{seurat object containing gene count and barcode locations (DefaultAssay = "Spatial")}

\item{\code{rgb}}{List containing the rgb code for each bead and each slice.}

\item{\code{nn}}{List of nearest neighbors in 2D space.}

\item{\code{territory}}{data.frame containing barcodes, cooridinates, RGB code
as well as territories and sub territories associated with each barcode.}

\item{\code{subTerritory}}{data.frame containing a subset of full data set.}

\item{\code{subCount}}{matrix containing a subset of full data set.}

\item{\code{subSO}}{seurat object containing a subset of full data set.}

\item{\code{combine}}{List containing territories and subterritories to extract and combine for downsteam analysis.}
}}

