% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vesalius}
\alias{vesalius}
\title{vesalius dummy data for ST analysis
\itemize{
  \item vesalius. Seurat object containing gene counts and coordinates
}}
\format{
Seurat Object
}
\usage{
data(vesalius)
}
\description{
vesalius dummy data for ST analysis
\itemize{
  \item vesalius. Seurat object containing gene counts and coordinates
}
}
\keyword{datasets}
