% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territory_identity_and_markers.R
\name{vesalius_deg_deseq2}
\alias{vesalius_deg_deseq2}
\title{DESeq negatove binomail for DEG}
\usage{
vesalius_deg_deseq2(seed, seed_id, query, query_id, params)
}
\arguments{
\item{seed}{count matrix for group 1}

\item{seed_id}{territory used in group 1}

\item{query}{count matrix for group 2}

\item{query_id}{territory used in group 2}

\item{params}{parameter value list (pval, log_fc, min_pct)}
}
\description{
DESeq negatove binomail for DEG
}
