% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{vesalius_kmeans}
\alias{vesalius_kmeans}
\title{kmeans segmentation function}
\usage{
vesalius_kmeans(
  vesalius,
  dimensions = seq(1, 3),
  col_depth = 10,
  embedding = "last",
  smooth_iter = 1,
  method = c("median", "iso", "box"),
  across_levels = "min",
  sigma = 1,
  box = 20,
  threshold = 0,
  neuman = TRUE,
  gaussian = TRUE,
  na.rm = FALSE,
  use_center = TRUE,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{vesalius}{a vesalius_assay}

\item{col_depth}{integer or vector of positive integers.
Colour depth used for segmentation. (see details)}

\item{smooth_iter}{integer - number of smoothing iterations}

\item{method}{character string for which method should be used for 
segmentation. Only "kmeans" supported.}

\item{across_levels}{character - method used to account for multiple 
smoothing levels. Select from: "min","mean", "max"}

\item{sigma}{numeric - standard deviation associated with isoblur (Gaussian)}

\item{box}{numeric describing box size (centered around a center pixel)
for smoothing}

\item{threshold}{numeric - discard pixels that are too low in value (cutoff
threshold only applied in box/median blurs).}

\item{neuman}{logical describing If Neumann boundary conditions should be
used, Dirichlet otherwise (default true, Neumann)}

\item{gaussian}{logical - use gaussian filter}

\item{na.rm}{logical describing if NA values should be removed}

\item{use_center}{logical - If TRUE, only the center pixel value will be used
during segmentation. If FALSE, all pixels will be used (see details)}

\item{cores}{numeric - number of cores that should be used}

\item{verbose}{logical - progress message output.}

\item{colDepth}{integer or vector of positive integers.
Colour depth used for segmentation.}
}
\value{
list containing segmented image as an active embedding and
territory cluster for all barcodes.
}
\description{
kmeans segmentation function
}
\details{
Run an interaive kmeans segmentation with the possibility 
to run multiple rounds of smoothing.
}
