% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{vesaliusObject-class}
\alias{vesaliusObject-class}
\title{The vesaliusObject class}
\description{
The vesaliusObject class is a high level obect that stores vesalius assay 
objects and provides an overview of all assays present. The purpose 
of this class is to provide a convenient way to store and process 
assays together. This object handles any type of spatial assay and alows
to run analysis on multiple assays simultaneously.
}
\section{Slots}{

\describe{
\item{\code{assays}}{list of \code{\link{vesalius_assay}}}

\item{\code{history}}{list containing overview of all analysis trials attempted
on each assay present.}
}}

