% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_mapping_metrics.R
\name{view_mapping_metrics}
\alias{view_mapping_metrics}
\title{visualize mapping metrics for vesalius assay}
\usage{
view_mapping_metrics(
  vesalius_assay,
  trial = "cost",
  cex_pt = 1,
  cex = 15,
  randomize = TRUE
)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}

\item{trial}{character string defining which metric trial to visualize}

\item{cex_pt}{numeric - point size for plotting}

\item{cex}{numeric - text size for plotting}

\item{randomize}{logical - whether to randomize point order for better visualization}
}
\value{
ggplot object showing mapping metrics
}
\description{
visualize mapping metrics for vesalius assay
}
