% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{align_index}
\alias{align_index}
\title{assign coordinates to matched indices}
\usage{
align_index(matched_index, coord, jitter = 0)
}
\arguments{
\item{matched_index}{data.frame containing matching pairs of
coordinates}

\item{coord}{data.frame containing coordinate data}

\item{jitter}{numeric - amount of jitter to add to duplicated coordinates}
}
\value{
adjusted coordinate data.frame where each point
receives the coordinates of its best match.
}
\description{
assign coordinates to matched indices
}
