% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_barcodes}
\alias{check_barcodes}
\title{checking overlap between barcodes in counts and coordinates}
\usage{
check_barcodes(counts, coordinates)
}
\arguments{
\item{counts}{character vector containing barcode names in count matrix}

\item{coordinates}{character vector containing barcode names in tile
data frame}
}
\value{
overlapping location between counts and coordinates
}
\description{
checking overlap between barcodes in counts and coordinates
}
\details{
Will throw in a warning if the overlap is not perfect.
}
