% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{check_image}
\alias{check_image}
\title{check if image is present}
\usage{
check_image(vesalius_assay)
}
\arguments{
\item{vesalius_assay}{vesalius assay object}
}
\value{
microscopy image
}
\description{
check if image is present
}
