% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{clean_trial}
\alias{clean_trial}
\title{clean trial annotation}
\usage{
clean_trial(trial)
}
\arguments{
\item{trial}{vector of strings with trial tag}
}
\value{
vector of strings with sanitize numeric value
}
\description{
clean trial annotation
}
