% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_dispatch.R
\name{cost_to_prob}
\alias{cost_to_prob}
\title{convert cost to prob}
\usage{
cost_to_prob(cost, n_cost)
}
\arguments{
\item{cost}{matrix containing cost}

\item{n_cost}{number of custom cost matrices parsed}
}
\value{
a probability matrix
}
\description{
convert cost to prob
}
