% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Territory_identity.R
\name{extractTerritories}
\alias{extractTerritories}
\title{extractTerritories provides a Seurat object from a Vesalius Territory}
\usage{
extractTerritories(
  image,
  seurat,
  seedID = NULL,
  morphologyFactor = 0,
  minBar = 10,
  verbose = TRUE,
  cores = 1
)
}
\arguments{
\item{image}{a data.frame containing territory information}

\item{seurat}{Seurat object. Contains all barcodes present on the spatial 
transcrtiptomic assay.}

\item{seedID}{integer or vector of integers describing territories that 
should be extracted.}

\item{morphologyFactor}{Integer or vector of integers describing growth or
shrink rate of territories. Only applied if seedID is provided.}

\item{minBar}{integer describing minimum number of barcodes in a territory 
set.}

\item{verbose}{logical - progress message output}

\item{cores}{numeric describing the number of cores used for the analysis.}
}
\value{
A Seurat oject containing barcodes taken from territory set.
}
\description{
extractTerritories provides a Seurat object from a Vesalius Territory
}
\details{
Once territories have been isolated from Vesalius images, the 
\code{extractTerritories} provides a convenient way to extract a set of 
territories, apply morphological operators, and return a Seurat object for 
further analysis. This Seurat Object can be analysed using Seurat's 
recommended pipeline.
}
\examples{
\dontrun{
data("Vesalius")
}

}
