% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{feature_dist}
\alias{feature_dist}
\title{compute the distance between seed and query signals}
\usage{
feature_dist(seed, query, norm = "noise", batch_size = 1000)
}
\arguments{
\item{seed}{seed signal}

\item{query}{query signal}
}
\value{
matrix with query as rows and seed as colmuns
}
\description{
compute the distance between seed and query signals
}
