% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{filter_tiles}
\alias{filter_tiles}
\title{filter tiles}
\usage{
filter_tiles(tesselation, coordinates, filter_threshold)
}
\arguments{
\item{tesselation}{data.frame output from the deldir function}

\item{coordinates}{data frame with original coordinates}

\item{filter_threshold}{numeric describing the quantile threshold value
to use for area filtering}
}
\value{
a list with 2 data frame. 1 with filtered tesselation results
2 filtered coordinate file.
}
\description{
filter tiles based on tile area and if they share an edge with the 
tesselation box
}
