% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_and_dispatch.R
\name{future_ves_to_cimg}
\alias{future_ves_to_cimg}
\title{convert ves embedding to image}
\usage{
future_ves_to_cimg(i, embeddings, dims, tiles)
}
\arguments{
\item{i}{index of embedding to use}

\item{embeddings}{matrix - embedding matrix}

\item{dims}{dimensions to use}

\item{tiles}{tile data frame used to reconstruct images}
}
\value{
cimg object of embedding
}
\description{
convert ves embedding to image
}
\details{
using this as a way to run this section in parallel
way to slow otherwise.
}
