% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metrics.R
\name{hclust_scores}
\alias{hclust_scores}
\title{hclust of mapping scores}
\usage{
hclust_scores(score, h, k, distance, verbose)
}
\arguments{
\item{score}{matrix containing mapping scores}

\item{h}{cutree height will take precedent over k}

\item{k}{number of cluster to get from hclust}

\item{distance}{which distance metric should be used to compute hclust}

\item{verbose}{logical for progress message output}
}
\value{
named vector - vector = clusters and names = barcodes
}
\description{
hclust of mapping scores
}
