% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate_assays}
\alias{integrate_assays}
\title{integrate counts from 2 vesalius assays}
\usage{
integrate_assays(
  mapped,
  reference,
  method = "CCAIntegration",
  nfeatures = 2000,
  signal = "variable_features",
  dimensions = 30,
  infer = TRUE,
  use_counts = "raw",
  cell_label_mapped = "Cells",
  cell_label_reference = "Cells",
  verbose = TRUE
)
}
\arguments{
\item{mapped}{vesalius_assay object - matched veslius assay (map_assays)}

\item{reference}{vesalius_assay object - reference vesalius_assay (seed)}

\item{method}{character - count integration method (methods provided by 
Seurat v5)}

\item{infer}{logical - back infer original counts by reversing reduced 
dimensional space roations.}

\item{use_counts}{character - which count matrix to use during integration}

\item{verbose}{logical - should progressed message be printed}
}
\description{
integrate counts from 2 vesalius assays
}
