% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{integrate_horizontally}
\alias{integrate_horizontally}
\title{Aling and integrate spatial assay from the same modality using super pixels}
\usage{
integrate_horizontally(
  seed_assay,
  query_assay,
  seed_trial = "last",
  query_trial = "last",
  scoring_method = "pearson",
  dimensions = seq(1, 30),
  scaling = 0.2,
  compactness = 1,
  n_centers = 2000,
  iter = 10000,
  index_selection = "random",
  threshold = 0.7,
  n_anchors = 20,
  depth = 1,
  allow_vertex_merge = FALSE,
  signal = "features",
  verbose = TRUE
)
}
\arguments{
\item{seed_assay}{vesalius_assay object - data to be mapped to}

\item{query_assay}{vesalius_assay objecy - data to map}

\item{seed_trial}{name of embedding to use for super pixel generation}

\item{query_trial}{name of embedding to use for super pixel generation}

\item{scoring_method}{method used to score the similarity between super 
pixels (pearson, spearman, kendall, coherence, index)}

\item{dimensions}{integer vector - which latent space dimensions should be
used for super pixel generation}

\item{scaling}{numeric ]0,1] describing image scale to consider 
during super pixel selection}

\item{compactness}{numeric ]0,Inf] - importance of the spatial component 
in relation to color similarity. See details}

\item{n_centers}{integer [3, max(spatial_index)] - number of super pixels
to generate in each image. See details}

\item{iter}{integer [1, Inf] - Number of iteration during graph matching
phase.}

\item{index_selection}{character (random, bubble) - how should initial super
pixel locations be selected. See detials}

\item{threshold}{numeric [0,1[ - similarity score threshold. Only super pixel
that score above this threshold will be used for graph matching}

\item{n_anchors}{integer [3, n_centers] - Number of graph anchors used during
graph matching}

\item{allow_vertex_merge}{logical - Should graph vertices be merged or should
them be repelled to new location. See details}

\item{signal}{character (features, counts, embeddings, "custom") - What should 
be used as cell signal for super pixel scoring. Seed details}

\item{verbose}{logical - should I be a noisy boy?}

\item{mut_extent}{numeric [0,1] - extent of alignment graph that can be 
subjected to mutations}

\item{mut_prob}{numeric [0,1] - probability of alignment graph mutation.}
}
\description{
Aling and integrate spatial assay from the same modality using super pixels
}
