% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{integrate_horizontally}
\alias{integrate_horizontally}
\title{Aling and integrate spatial assay from the same modality using super pixels}
\usage{
integrate_horizontally(
  seed_assay,
  query_assay,
  k = 20,
  dimensions = seq(1, 30),
  mapping = "div",
  signal = "variable_features",
  use_norm = "raw",
  verbose = TRUE
)
}
\arguments{
\item{seed_assay}{vesalius_assay object - data to be mapped to}

\item{query_assay}{vesalius_assay objecy - data to map}

\item{k}{int ]2, n_points] number of neareset neighbors to be considered for
neighborhodd computation.}

\item{mapping}{character string (div - exact) - mapping strategy. 
Divide and conquer (approximate) or Exact mapping.}

\item{signal}{character (features, counts, embeddings, "custom") - What should 
be used as cell signal for super pixel scoring. Seed details}

\item{use_norm}{character - which count data to use}

\item{verbose}{logical - should I be a noisy boy?}
}
\description{
Aling and integrate spatial assay from the same modality using super pixels
}
\details{
Hex grid is expanded to include all points and then reduced to exclude
exmpty triangles.
}
