% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{kmeans_segmentation}
\alias{kmeans_segmentation}
\title{kmeans segmentation function}
\usage{
kmeans_segmentation(
  vesalius_assay,
  dimensions = seq(1, 3),
  col_resolution = 10,
  embedding = "last",
  use_center = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{vesalius_assay}{a vesalius_assay}

\item{dimensions}{numeric vector of latent space dimensions to use.}

\item{col_resolution}{integer or vector of positive integers.
Colour depth used for segmentation.}

\item{embedding}{character string describing which embedding should
be used.}

\item{use_center}{logical - If TRUE, only the center pixel value will be used
during segmentation. If FALSE, all pixels will be used (see details)}

\item{verbose}{logical - progress message output.}
}
\value{
list containing segmented image as an active embedding and
territory cluster for all barcodes.
}
\description{
kmeans segmentation function
}
\details{
Run an interaive kmeans segmentation with the possibility 
to run multiple rounds of smoothing.
}
