% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology_operators.R
\name{layer_territory}
\alias{layer_territory}
\title{layer_territory generates layer from the outside to the inside of 
a territory}
\usage{
layer_territory(
  vesalius_assay,
  territory = NULL,
  trial = "last",
  layer_depth = NULL,
  morphology_factor = 0,
  verbose = TRUE
)
}
\arguments{
\item{vesalius_assay}{vesalius_assay object}

\item{territory}{integer or vector of integers desrcining territories 
to morph.}

\item{trial}{character string - which territory trial that 
should be used to select
territorires. Default is last one computed}

\item{layer_depth}{integer describing the number of final layers.}

\item{morphology_factor}{integer or vector of integers describing growth
and/or shrink extent.}

\item{verbose}{logical - progress message output.}
}
\value{
a vesalius_assay
}
\description{
layer_territory generates layer from the outside to the inside of 
a territory
}
\details{
Each territory can be subdivided into a series of layers. 
Each layer will be considered a seperate territory and can be treated as
such for functions such as \code{\link{identify_markers}} and 
\code{\link{territory_plot}}.

However, all other territories present will be labled as "out". 
This means that for the time being you can only work with a single
territory at a time.
}
\examples{
\dontrun{
data(vesalius)
# First we build a simple object
ves <- build_vesalius_object(coordinates, counts)
# We can do a simple run
ves <- build_vesalius_embeddings(ves)

# simple smoothing
ves <- smooth_image(ves, dimensions = seq(1, 30))

# quick segmentation
ves <- segment_image(ves, dimensions = seq(1, 30))

# isolate territories
ves <- isolate_territories(ves)

# morph territory

ves <- layer_territory(ves)

# view territory morphing
territory_plot(ves)
}
}
