% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metrics.R
\name{leiden_scores}
\alias{leiden_scores}
\title{leiden clustering applied to mapping scores}
\usage{
leiden_scores(score, resolution, nn, verbose)
}
\arguments{
\item{resolution}{numeric - clustering resolution}

\item{nn}{integer - number of nearest neighbors based on scores}

\item{verbose}{logical - if progress message should be printed}

\item{scores}{matrix containing mapping scores}
}
\value{
named vector - vector = clusters and names = barcodes
}
\description{
leiden clustering applied to mapping scores
}
\details{
Unweighted leiden custering could be weighted by score
this could help discriminate of scores are low but still above 
filter criteria
}
