% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{match_index}
\alias{match_index}
\title{Finding optimal mapping between query and seed using a mini batch 
approach}
\usage{
match_index(idx, cost_matrix)
}
\arguments{
\item{idx}{integer - index of batch}

\item{cost_matrix}{list of matrices with cost between query (rows) and 
seed (columns)}
}
\value{
data.frame with point in query (from) mapped to 
which point in seed (to) as well as cost of mapping for that pair
}
\description{
Finding optimal mapping between query and seed using a mini batch 
approach
}
\details{
Uses a Hungarian solver to minimize overall cost of 
pair assignement.
}
