% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{niche_composition}
\alias{niche_composition}
\title{Method dispatch function for neighborhood selection - added
flavor specific to composition}
\usage{
niche_composition(
  coord,
  vesalius_assay,
  method,
  cell_label = NULL,
  k = 20,
  depth = 3,
  radius = 20
)
}
\arguments{
\item{coord}{data.frame - coordinates of assay (barcodes, x, y)}

\item{vesalius_assay}{vesalius_assay object containing cell labels}

\item{method}{character - which method should be use to collect 
neighborhood - switch matches}

\item{cell_label}{character - which column contains the cell label names?}

\item{k}{int - how many nearest neighbors from KNN algorithm}

\item{depth}{int - graph path depth to define neighborhood 
0 = self, 1 = direct neigbors, 2 = neighbors of neighbors, etc}

\item{radius}{- numeric - radius around center cell}
}
\value{
a composition matrix (cell_labels X cell barcodes)
}
\description{
Method dispatch function for neighborhood selection - added
flavor specific to composition
}
