% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{pearson_approx}
\alias{pearson_approx}
\title{compute fast pearson correlation between matrices}
\usage{
pearson_approx(seed, query)
}
\arguments{
\item{seed}{matrix/sparse matrix of seed cells}

\item{query}{matrix/sparse matrix of query cells}
}
\value{
correlation matrix between all cells in seed and query
}
\description{
compute fast pearson correlation between matrices
}
