% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{point_mapping}
\alias{point_mapping}
\title{mapping points between data sets}
\usage{
point_mapping(
  query_signal,
  query_assay,
  cost,
  seed_signal,
  seed_assay,
  method = "pearson",
  neighborhood = "knn",
  k = 20,
  radius = 0.05,
  depth = 1,
  batch_size = 10000,
  epochs = 1,
  use_cost = c("feature", "niche"),
  threshold = 0.5,
  filter_cells = FALSE,
  seed_territory_labels = "Territory",
  query_territory_labels = "Territory",
  seed_meta_labels = NULL,
  query_meta_labels = NULL,
  digits = 4,
  verbose = TRUE
)
}
\arguments{
\item{query_signal}{processed query signal from query assay}

\item{query_assay}{vesalius_assay object}

\item{cost}{matrix - matrix of size n (query cells) by p (seed cells)
containing custom cost matrix.}

\item{seed_signal}{processed seed signal from seed assay}

\item{seed_assay}{vesalius_assay object}

\item{method}{character - correlation method}

\item{neighborhood}{character - neighborhood method}

\item{k}{int size of niche (knn)}

\item{radius}{0.05 proportion of max distance to use as radius for 
neighborhood}

\item{depth}{graph path depth to condsider for neighborhood.}

\item{batch_size}{int number of points in each query batch}

\item{epochs}{numeric - number of epochs}

\item{use_cost}{character string defining how should total cost be computer
Available: feature, niche, territory, composition (See details for combinations}

\item{threshold}{score threshold below which indicices should be removed.
Scores will always be between 0 and 1}

\item{filter_cells}{logical - filter cells}

\item{seed_territory_labels}{character - seed territory labels}

\item{query_territory_labels}{character - query territory labels}

\item{seed_meta_labels}{character - seed meta labels}

\item{query_meta_labels}{character - query meta labels}

\item{digits}{numeric - number of digits}

\item{verbose}{logical - out progress to console}
}
\value{
list of matched and aligned coordinates in query
}
\description{
mapping points between data sets
}
