% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{point_mapping}
\alias{point_mapping}
\title{mapping points between data sets}
\usage{
point_mapping(
  query_signal,
  query_assay,
  cost,
  seed_signal,
  seed_assay,
  method = "pearson",
  neighborhood = "knn",
  k = 20,
  radius = 0.05,
  depth = 1,
  batch_size = 10000,
  epochs = 1,
  use_cost = c("feature", "niche"),
  threshold = 0.5,
  filter_cells = FALSE,
  seed_territory_labels = "Territory",
  query_territory_labels = "Territory",
  seed_meta_labels = NULL,
  query_meta_labels = NULL,
  digits = 4,
  verbose = TRUE
)
}
\arguments{
\item{query_signal}{processed query signal from query assay}

\item{query_assay}{vesalius_assay object}

\item{cost}{matrix - matrix of size n (query cells) by p (seed cells)
containing custom cost matrix.}

\item{seed_signal}{processed seed signal from seed assay}

\item{seed_assay}{vesalius_assay object}

\item{method}{character - which pearson correlation to use? 
(pearson, pearson_fast, pearson_exact) TO BE DEPRECIATED}

\item{neighborhood}{character - how should the neighborhood be selected?
"knn", "radius", "graph"}

\item{k}{int ]2, n_points] number of neareset neighbors to be considered for
neighborhodd computation.}

\item{radius}{numeric ]0,1[ proportion of max distance between points 
to consider for the neighborhood}

\item{depth}{int [1, NA] graph depth from cell to consider from neighborhood
(See details)}

\item{batch_size}{number of points per batch in query during assignment
problem solving}

\item{epochs}{int [1, [ - Across how many epochs should the mapping occur?
Note that if batch_size is larger than both data sets, running across epochs
will not improve mapping.}

\item{use_cost}{character string defining how should total cost be computer
Available: feature, niche, territory, composition, cell_type (See details for combinations
and custom matrices)}

\item{threshold}{numeric - Any cell (accros all matrices except cell_type) whose max score
falls below this threshold will be removed.}

\item{filter_cells}{logical - If cell_type cost matrix is requested, remove cells from
query that are not present in the seed data set.}

\item{seed_territory_labels}{character - name of column in seed containing territories
which should be used for territory cost}

\item{query_territory_labels}{character - name of column in query containing territories
which should be used for territory cost}

\item{seed_meta_labels}{character - name of columns in seed containing meta information
to be parse to the mapped object. First name will be used for cell_type cost if
requested.}

\item{query_meta_labels}{character - name of columns in query containing meta information
to be parse to the mapped object. First name will be used for cell_type cost if
requested.}

\item{digits}{number of digits to retain after computing correlation. More digits
equals to a more precise mapping but significantly increase run time.}

\item{verbose}{logical - out progress to console}
}
\value{
list containing matched coordinates, cost matrices, and cost by epoch
}
\description{
mapping points between data sets
}
