% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_assays.R
\name{point_mapping}
\alias{point_mapping}
\title{mapping points between data sets}
\usage{
point_mapping(
  seed,
  seed_signal,
  query,
  query_signal,
  mapping = "div",
  k = 20,
  batch_size = 1000,
  custom_cost = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{seed}{vesalius_assay object}

\item{seed_signal}{processed seed signal from seed assay}

\item{query}{vesalius_assay object}

\item{query_signal}{processed query signal from query assay}

\item{mapping}{exact mapping or div (divide and conquer)}

\item{k}{int size of niche (knn)}

\item{batch_size}{int number of points in each query batch}

\item{custom_cost}{matrix - matrix of size n (query cells) by p (seed cells)
containing custom cost matrix. Used instead of vesalius cost matrix}

\item{overwrite}{logical - if custom_cost is not NULL, should this matrix
be added to the vesalius matrix or should it overwrite it completely.}

\item{verbose}{logical - out progress to console}
}
\value{
list of matched and aligned coordinates in query
}
\description{
mapping points between data sets
}
