% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_assays.R
\name{point_mapping}
\alias{point_mapping}
\title{mapping points between data sets}
\usage{
point_mapping(
  query_signal,
  query_assay,
  cost,
  seed_signal,
  seed_assay,
  neighborhood = "knn",
  k = 20,
  radius = 0.05,
  depth = 1,
  batch_size = 1000,
  use_cost = c("feature", "niche"),
  threshold = 0.5,
  verbose = TRUE
)
}
\arguments{
\item{query_signal}{processed query signal from query assay}

\item{cost}{matrix - matrix of size n (query cells) by p (seed cells)
containing custom cost matrix.}

\item{seed_signal}{processed seed signal from seed assay}

\item{k}{int size of niche (knn)}

\item{radius}{0.05 proportion of max distance to use as radius for 
neighborhood}

\item{depth}{graph path depth to condsider for neighborhood.}

\item{batch_size}{int number of points in each query batch}

\item{use_cost}{character string defining how should total cost be computer
Available: feature, niche, territory, composition (See details for combinations}

\item{threshold}{score threshold below which indicices should be removed.
Scores will always be between 0 and 1}

\item{verbose}{logical - out progress to console}

\item{query}{vesalius_assay object}

\item{seed}{vesalius_assay object}
}
\value{
list of matched and aligned coordinates in query
}
\description{
mapping points between data sets
}
