% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_and_tiles.R
\name{raw_norm}
\alias{raw_norm}
\title{raw norm}
\usage{
raw_norm(counts, use_count = "raw")
}
\arguments{
\item{counts}{seurat object containing counts}
}
\value{
list with seurat object used later and raw counts to be stored in
the vesalius objects
}
\description{
no normalisation applied simply return raw counts
}
