% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_pixel.R
\name{select_initial_indices}
\alias{select_initial_indices}
\title{select inital indices}
\usage{
select_initial_indices(
  coordinates,
  embeddings,
  n_centers = 500,
  max_iter = 500
)
}
\arguments{
\item{coordinates}{data frame containing spatial coordinates of beads}

\item{embeddings}{matrix containing embedding values - full pixel image}

\item{n_centers}{numeric number of beads to select as super pixel centers}

\item{max_iter}{numeric number of iteration before returning result if 
no coveregnce.}
}
\value{
barcodes of starting coordinates
}
\description{
select inital indices
}
