% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viewLayerExpression}
\alias{viewLayerExpression}
\title{viewGeneExpression - plot gene expression in layered Vesalius territories}
\usage{
viewLayerExpression(image, counts, genes = NULL, normalise = TRUE, cex = 10)
}
\arguments{
\item{image}{a Vesalius data frame containing barcodes, x, y, cc, value,
cluster, and territory.}

\item{counts}{count matrix - either matrix, sparse matrix or seurat object
This matrix should contain genes as rownames and cells/barcodes as colnames}

\item{genes}{character - gene of interest (only on gene at a time)}

\item{normalise}{logical - If TRUE, gene expression values will be min/max
normalised.}

\item{cex}{numeric - font size modulator}
}
\value{
a ggplot object (geom_raster)
}
\description{
viewGeneExpression - plot gene expression in layered Vesalius territories
}
\details{
Visualization of gene expression a layered territory.
After isolation or a territory, Vesalius can apply territory layering. The
expression shown here describes the mean expression in each layer.

Gene expression is shown "as is". This means that if no transformation
is applied to the data then normalized raw count will be used for each layer.

If normalization and scaling have been applied, normalized counts will be
used for each layer

This also applies to any data transformation applied on the count matrix
or the Seurat object.

We recommend using Seurat Objects.

NOTE : You might be promoted to use geom_tiles instead of geom_raster.
However - this will increase file size to unreasonable heights.
}
\examples{
\dontrun{
data(Vesalius)
}
}
